### 2020-06-07 #2299 @kyleniemeyer
@rajeshrinet regarding the paper build error, I think the issue is the SVG badge you are trying to include. Can you remove, or convert to a PNG (for example)?
