### 2020-06-07 dd7650d
Update paper.md
