### 2020-06-07 5372292
reduce bib size
