+++ docs/paper/paper.bib
@@ -1,2531 +1,3 @@
-@book{landau1987fluid,
-    title={Fluid mechanics, 2nd},
-    author={Landau, L. D. and Lifshitz, E. M.},
-    year={1987},
-    publisher={London: Elsevier}
-}
-@article{bolitho2020,
-author = {Bolitho, Austen and Singh, Rajesh and Adhikari, R.},
-journal = {Phys. Rev. Lett.},
-doi={},
-pages = {088003},
-volume = {124},
-number = {8},
-year = {2020}
-}
-
-@book{maxwell1881treatise,
-  title={A treatise on electricity and magnetism},
-  author={Maxwell, J. C.},
-  volume={1},
-  year={1873},
-  pages={89},
-  publisher={Clarendon press, Oxford}
-}
-
-@book{landau1959fluid,
-    title={Fluid mechanics},
-    author={Landau, L. D. and Lifshitz, E. M.},
-    volume={6},
-    year={1959},
-    url={https://archive.org/details/FluidMechanics},
-    publisher={Pergamon Press, New York}
-}
-
-@book{mackay2003information,
-  title={Information theory, inference and learning algorithms},
-  author={MacKay, D. J. C.},
-  year={2003},
-  publisher={Cambridge University Press}
-}
-
-@book{sivia2006data,
-  author    = {Sivia, D. and Skilling, J.},
-  title     = {Data analysis: a {B}ayesian tutorial},
-  publisher = {Oxford University Press, Oxford},
-  year      = {2006},
-}
-
-@book{landau2013electrodynamics,
-    title={Electrodynamics of continuous media},
-    author={Landau, L. D. and Lifshitz, E. M.},
-    year={1987},
-    url={https://archive.org/details/ElectrodynamicsOfContinuousMedia},
-    publisher={Pergamon, Oxford}
-}
-
-@book{goldstein2011classical,
-    title={Classical mechanics},
-    author={Goldstein, H.},
-    year={1980},
-    url={https://books.google.co.in/books?isbn=8131758915},
-    publisher={Addison-Wesley}
-}
-
-@book{batchelor2000introduction,
-    title={An introduction to fluid dynamics},
-    author={Batchelor, G.~K.},
-    year={2000},
-    url={https://books.google.co.in/books?isbn=0521663962},
-    publisher={Cambridge university press}
-}
-
-@book{perrin2013brownian,
-    title={Brownian movement and molecular reality},
-    author={Perrin, J.},
-    year={1910},
-    url={https://books.google.co.in/books?isbn=0486174727},
-    publisher={Taylor \& Francis}
-}
-
-@article{lorentz1896eene,
-    author={Lorentz, H. A.},
-    title={A general theorem concerning the motion of a viscous fluid and a few consequences derived from it},
-    volume={5},
-    pages={168--175},
-    year={1896},
-    journal={Versl. Konigl. Akad. Wetensch. Amst},
-    publisher={ Zittingsverlag Akad.\ v.\ Wet.}
-}
-
-@book{gardiner1985handbook,
-    title={Handbook of stochastic methods},
-    author={Gardiner, C. W.},
-    year={1985},
-    url={http://www.springer.com/gb/book/9783540707127},
-    publisher={Springer Berlin}
-}
-
-@book{kleman2007soft,
-  title={Soft matter physics: an introduction},
-  author={Kleman, M. and Laverntovich, O. D.},
-  year={2007},
-  url={https://link.springer.com/book/10.1007%2Fb97416},
-  publisher={Springer Science \& Business Media}
-}
-
-@book{anderson1987broken,
-    author={P. W. Anderson and D. L. Stein},
-    year={1987},
[…]
