+++ docs/paper/paper.md
@@ -37,7 +37,7 @@ Our method relies on the reduction of linear elliptic partial differential equat
 
 ![\label{fig:example}](figure.png)
 
-The first step is the representation of the solution of an elliptic partial differential equation (PDE) in a three-dimensional volume $V$ as an integral over the boundary of the surface $S$ [@fkg1930bandwertaufgaben, @ladyzhenskaya1969, @youngren1975stokes, @zick1982stokes, @pozrikidis1992, @muldowney1995spectral, @cheng2005heritage, @singh2015many]. For the Laplace equation, this is the classical theorem of Green [@jackson1962classical]; for the Stokes equation, it is the generalization obtained by Lorentz [@fkg1930bandwertaufgaben, @lorentz1896eene, @ladyzhenskaya1969]. The integral representation leads to a linear integral equation that provides a functional relation between the field and its flux on S. Thus, if the surface flux in the Laplace equation is specified, the surface concentration is determined by the solution of the Laplace boundary integral equation. Similarly, if the surface velocity in the Stokes equation is specified, the surface traction is determined by the solution of the Stokes boundary integral equation. This transformation of the governing PDEs is the most direct way of relating boundary conditions (surface flux, slip velocities) to boundary values (surface concentration, surface traction). It reduces the dimensionality of the problem from a three-dimensional one in V to a two-dimensional one on S. The second step is the spectral expansion of the field and its flux in terms of global basis functions on S. We use the geometry-adapted tensorial spherical harmonics, which provide an unified way of expanding both scalar and vector quantities on the surface of a sphere. These functions are both complete and orthogonal and provide representations of the three-dimensional rotation group [@hess2015tensors]. Thus, symmetries of the active boundary conditions can be represented in a straightforward and transparent manner. The third step is the discretization of the integral equation using the procedure of Ritz and Galerkin [@boyd2001chebyshev, @finlayson1966method], which reduces it to an infinite-dimensional self-adjoint linear system in the expansion coefficients. This exploits the orthogonality of the basis functions on the sphere. The matrix elements of the linear system can be evaluated analytically in terms of the Green's functions of the respective elliptic equations. The fourth step is the truncation of the infinite-dimensional linear system to a finite-dimensional one that can be solved by standard methods of linear algebra adapted for self-adjoint systems [@saad2003iterative]. Analytical solution can be obtained by Jacobi iteration, which is equivalent to Smoluchowski's method of reflections. Numerical solutions can be obtained by the conjugate gradient method, at a cost quadratic in the number of unknowns. From this solution, we can reconstruct the field and the flux on the boundary, use these to determine the fields in the bulk, and from there, compute derived quantities. These steps have been elaborated in several papers [@ghose2014irreducible, @singh2015many, @singh2016crystallization, @singh2018generalized, @singh2019competing] and we do not repeat them in detail here. 
[…]
