### 2020-06-07 #2299 @rajeshrinet
> @rajeshrinet regarding the paper build error, I think the issue is the SVG badge you are trying to include. Can you remove, or convert to a PNG (for example)?

@kyleniemeyer thanks very much for your comments. I have removed the badge and pared down the bib file. I hope that solves the PDF compilation.
