### 2020-06-07 #2299 @rajeshrinet
@kyleniemeyer it seems I need use [@article1] for each citation. It can not be done in a bunch like [@article1, @article2]?
