### 2020-06-07 #2299 @kyleniemeyer
@rajeshrinet You can include multiple citations together, but you need semicolons rather than commas between each citation: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
