+++ docs/paper/paper.md
@@ -62,6 +62,6 @@ for different boundary conditions. Commonly used force fields in colloidal syste
 
 # Acknowledgements
 
-RS is funded by a Royal Society-SERB Newton International Fellowship. RA thanks the Isaac Newton Trust for an Early Career Support grant. We would like to thank collaborators and colleagues, in both theory and experiment, for numerous discussions that have enriched our understanding of hydrodynamic and phoretic phenomena. In alphabetical order, they are: Ayan Banerjee, Mike Cates, Shyam Date, Aleks Donev, Erika Eiser, Daan Frenkel, Somdeb Ghose, Ray Goldstein, John Hinch, Abhrajit Laskar, Tony Ladd, Raj Kumar Manna, Ignacio Pagonabarraga, Dave Pine, Thalappil Pradeep, Rajiah Simon, Howard Stone, Ganesh Subramanian, P. B. Sunil Kumar, and Shashi Thutupalli. We acknowledge design ideas and code contributions from Rajeev Singh and Abhrajit Laskar in the initial stages of development. This work was funded in part by the European Research Council under the EU’s Horizon 2020 Program.
+We would like to thank collaborators and colleagues, in both theory and experiment, for numerous discussions that have enriched our understanding of hydrodynamic and phoretic phenomena. In alphabetical order, they are: Ayan Banerjee, Mike Cates, Shyam Date, Aleks Donev, Erika Eiser, Daan Frenkel, Somdeb Ghose, Ray Goldstein, John Hinch, Abhrajit Laskar, Tony Ladd, Raj Kumar Manna, Ignacio Pagonabarraga, Dave Pine, Thalappil Pradeep, Rajiah Simon, Howard Stone, Ganesh Subramanian, P. B. Sunil Kumar, and Shashi Thutupalli. We acknowledge design ideas and code contributions from Rajeev Singh and Abhrajit Laskar in the initial stages of development. This work was funded in part by the European Research Council under the EU’s Horizon 2020 Program, Royal Society-SERB Newton International Fellowship to RS, and an Early Career Grant to RA from the Isaac Newton Trust.
 
 # References
