+++ docs/paper/paper.bib
@@ -258,3 +258,12 @@
     doi={10.1016/j.enganabound.2004.12.001},
     publisher={Elsevier}
 }
+
+@article{singh2019PyStokes,
+	author = "{R. Singh} and Adhikari, R.",
+	journal = "arXiv:1910.00909",
+	title = "{Hydrodynamic and phoretic interactions of active particles in Python}",
+	url = "https://arxiv.org/abs/1910.00909",
+	year = "2019"
+}
+
