+++ docs/paper/paper.bib
@@ -1,15 +1,4 @@
 
-@article{ghose2014irreducible,
-    title={Irreducible Representations of Oscillatory and Swirling Flows in Active Soft Matter},
-    author={S. Ghose and Adhikari, R.},
-    journal={Phys. Rev. Lett.},
-    volume={112},
-    number={11},
-    pages={118102},
-    year={2014},
-    doi={10.1103/PhysRevLett.112.118102},
-    publisher={APS}
-}
 @book{saad2003iterative,
     title={Iterative methods for sparse linear systems},
     author={Saad, Y.},
@@ -17,7 +6,6 @@
     doi={10.1137/1.9780898718003.bm},
     publisher={SIAM}
 }
-
 
 @article{finlayson1966method,
   title={The method of weighted residuals - a review},
@@ -84,7 +72,7 @@
 }
 @article{singh2019competing,
   title={Competing chemical and hydrodynamic interactions in autophoretic colloidal suspensions},
-  author={{R. Singh} and Adhikari, R. and Cates, M. E.},
+  author={Singh, R and Adhikari, R. and Cates, M. E.},
   journal={J. Chem. Phys.},
   volume={151},
   pages={044901},
@@ -195,9 +183,9 @@
 }
 
 @article{bolitho2020,
-    author = {Bolitho, Austen and Singh, Rajesh and Adhikari, R.},
+    author = {Bolitho, A. and Singh, R. and Adhikari, R.},
     journal = {Phys. Rev. Lett.},
-    doi={},
+    doi={10.1103/PhysRevLett.124.088003},
     pages = {088003},
     volume = {124},
     number = {8},
@@ -260,10 +248,21 @@
 }
 
 @article{singh2019PyStokes,
-	author = "{R. Singh} and Adhikari, R.",
-	journal = "arXiv:1910.00909",
-	title = "{Hydrodynamic and phoretic interactions of active particles in Python}",
-	url = "https://arxiv.org/abs/1910.00909",
-	year = "2019"
+    author = "{Singh, R. and Adhikari, R.",
+    journal = "arXiv:1910.00909",
+    title = "{Hydrodynamic and phoretic interactions of active particles in Python}",
+    url = "https://arxiv.org/abs/1910.00909",
+    year = "2019"
 }
 
+@article{thutupalli2018FIPS,
+	author = "Thutupalli, S. and Geyer, D. and Singh, R. and Adhikari, R. and Stone, H. A.",
+	doi = "10.1073/pnas.1718807115",
+	journal = "Proc. Natl. Acad. Sci.",
+	number = "21",
+	pages = "5403--5408",
+	title = "{Flow-induced phase separation of active particles is controlled by boundary conditions}",
+	volume = "115",
+	year = "2018",
+	year2 = "2018"
+}
