+++ docs/paper/paper.bib
@@ -1,4 +1,3 @@
-
 @book{saad2003iterative,
     title={Iterative methods for sparse linear systems},
     author={Saad, Y.},
@@ -8,22 +7,21 @@
 }
 
 @article{finlayson1966method,
-  title={The method of weighted residuals - a review},
-  author={Finlayson, B. A. and Scriven, L. E.},
-  journal={Appl. Mech. Rev},
-  volume={19},
-  number={9},
-  pages={735--748},
-  year={1966}
+    title={The method of weighted residuals - a review},
+    author={Finlayson, B. A. and Scriven, L. E.},
+    journal={Appl. Mech. Rev},
+    volume={19},
+    number={9},
+    pages={735--748},
+    year={1966}
 }
 
-
 @book{boyd2001chebyshev,
-  title={Chebyshev and Fourier spectral methods},
-  author={Boyd, J. P.},
-  year={2000},
-  url={https://store.doverpublications.com/0486411834.html},
-  publisher={Dover}
+    title={Chebyshev and Fourier spectral methods},
+    author={Boyd, J. P.},
+    year={2000},
+    url={https://store.doverpublications.com/0486411834.html},
+    publisher={Dover}
 }
 
 
@@ -35,7 +33,6 @@
     publisher={Wiley}
 }
 
-
 @book{hess2015tensors,
     title={Tensors for physics},
     author={Hess, S.},
@@ -43,6 +40,7 @@
     doi={10.1007/978-3-319-12787-3},
     publisher={Springer}
 }
+
 @book{ladyzhenskaya1969,
     title={The mathematical theory of viscous incompressible flow},
     author={Ladyzhenskaia, O. A.},
@@ -52,6 +50,7 @@
     url={https://books.google.co.in/books?id=qVXvAAAAMAAJ},
     publisher={Gordon and Breach}
 }
+
 @article{lorentz1896eene,
     author={Lorentz, H. A.},
     title={A general theorem concerning the motion of a viscous fluid and a few consequences derived from it},
@@ -61,6 +60,7 @@
     journal={Versl. Konigl. Akad. Wetensch. Amst},
     publisher={ Zittingsverlag Akad.\ v.\ Wet.}
 }
+
 @article{fkg1930bandwertaufgaben,
     title={{\"U}ber die Bandwertaufgaben der Hydrodynamik z{\"a}her Fl{\"u}ssig-keiten},
     author={Odqvist, F. K. G.},
@@ -70,18 +70,19 @@
     doi={10.1007/BF01194638},
     year={1930}
 }
+
 @article{singh2019competing,
-  title={Competing chemical and hydrodynamic interactions in autophoretic colloidal suspensions},
-  author={Singh, R and Adhikari, R. and Cates, M. E.},
-  journal={J. Chem. Phys.},
-  volume={151},
-  pages={044901},
-  doi={10.1063/1.5090179},
-  year={2019}
+    title={Competing chemical and hydrodynamic interactions in autophoretic colloidal suspensions},
+    author={Singh, R and Adhikari, R. and Cates, M. E.},
+    journal={J. Chem. Phys.},
+    volume={151},
+    pages={044901},
+    doi={10.1063/1.5090179},
+    year={2019}
 }
  
 @article{singh2018generalized,
-    author={R. Singh    and R. Adhikari},
+    author={R. Singh and R. Adhikari},
     title={Generalized {S}tokes laws for active colloids and their applications},
     journal={J. Phys. Commun.},
     volume={2},
@@ -203,7 +204,6 @@
     doi={10.1017/S0022112075001486},
     publisher={Cambridge Univ Press}
 }
-
 
 @article{zick1982stokes,
     title={Stokes flow through periodic arrays of spheres},
