+++ docs/paper/paper.bib
@@ -248,7 +248,7 @@
 }
 
 @article{singh2019PyStokes,
-    author = "{Singh, R. and Adhikari, R.",
+    author ="Singh, R. and Adhikari, R.",
     journal = "arXiv:1910.00909",
     title = "{Hydrodynamic and phoretic interactions of active particles in Python}",
     url = "https://arxiv.org/abs/1910.00909",
@@ -256,13 +256,13 @@
 }
 
 @article{thutupalli2018FIPS,
-	author = "Thutupalli, S. and Geyer, D. and Singh, R. and Adhikari, R. and Stone, H. A.",
-	doi = "10.1073/pnas.1718807115",
-	journal = "Proc. Natl. Acad. Sci.",
-	number = "21",
-	pages = "5403--5408",
-	title = "{Flow-induced phase separation of active particles is controlled by boundary conditions}",
-	volume = "115",
-	year = "2018",
-	year2 = "2018"
+	title = {Flow-induced phase separation of active particles is controlled by boundary conditions},
+    author={S. Thutupalli and D. Geyer and R. Singh and R. Adhikari and H. A. Stone},
+	journal={Proc. Natl. Acad. Sci.},
+	volume = {115},
+	number = {21},
+	pages = {5403--5408},
+	year = {2018},
+	doi={10.1073/pnas.1718807115},
+	year={2018},
 }
