+++ docs/changelogs.md
@@ -1,0 +1,13 @@
+# PyStokes changelogs
+
+## 2.1.1 
+* Second release of PyStokes
+
+## 2.0.0 
+* PyStokes has been merged with all of PyLaplace and PyForces
+
+## 1.0.0 
+* Add files from GitLab (where it was developed) for PyStokes, PyLaplace, and PyForces. 
+
+## 0.1 
+* First pre-release of PyStokes 
