### 2020-06-08 #2299 @arfon
@rajeshrinet - also, it would be very helpful if you could truncate your BibTeX file to only include the actual papers/entities that you're citing in the JOSS paper.
