### 2020-06-08 f685bac
remove all entries non-cited
