+++ docs/paper/paper.bib
@@ -92,21 +92,6 @@
     year={2018}
 }
 
-@article{singh2018fast,
-    title={Fast Bayesian inference of the multivariate Ornstein-Uhlenbeck process},
-    author={Singh, R. and Ghosh, D. and Adhikari, R.},
-    journal={Phys. Rev. E},
-    volume={98},
-    issue={1},
-    pages={012136},
-    numpages={9},
-    year={2018},
-    month={Jul},
-    publisher={American Physical Society},
-    doi={10.1103/PhysRevE.98.012136},
-    url={https://link.aps.org/doi/10.1103/PhysRevE.98.012136}
-}
-
 @article{singh2017fluctuation,
     author={R. Singh    and R. Adhikari},
     title={Fluctuating hydrodynamics and the {B}rownian motion of an active colloid near a wall},
@@ -176,13 +161,6 @@
     publisher={Annual Reviews 4139 El Camino Way, PO Box 10139, Palo Alto, CA 94303-0139, USA}
 }
 
-@book{landau1987fluid,
-    title={Fluid mechanics, 2nd},
-    author={Landau, L. D. and Lifshitz, E. M.},
-    year={1987},
-    publisher={London: Elsevier}
-}
-
 @article{bolitho2020,
     author = {Bolitho, A. and Singh, R. and Adhikari, R.},
     journal = {Phys. Rev. Lett.},
@@ -247,7 +225,7 @@
     publisher={Elsevier}
 }
 
-@article{singh2019PyStokes,
+@article{singh2019Hydrodynamic,
     author ="Singh, R. and Adhikari, R.",
     journal = "arXiv:1910.00909",
     title = "{Hydrodynamic and phoretic interactions of active particles in Python}",
