### 2020-06-08 #2299 @rajeshrinet
@arfon I have now removed all the remaining entries from bib which were not cited in JOSS paper.
