### 2020-06-08 e3deb3d
docstrings in mima
