+++ pystokes/mima.pyx
@@ -1,11 +1,3 @@
-"""
-* Mollified irreducible multipole approach (MIMA)
-* MIMA resolves the fluid flow and solves Stokes equation directly.
-* For further details of this method, read chapter 10 of the thesis: 
-  https://dspace.imsc.res.in/xmlui/handle/123456789/418 
-  or at https://www.imsc.res.in/xmlui/handle/123456789/418
-"""
-
 cimport cython
 from libc.math cimport sqrt, exp, pow, erfc, sin, cos
 from cython.parallel import prange
@@ -21,6 +13,13 @@ cdef extern from "complex.h" nogil:
 @cython.nonecheck(False)
 @cython.wraparound(False)
 cdef class Mima3D:
+    """
+    Mollified irreducible multipole approach (MIMA) of a 3D Stokes flow
+    MIMA resolves the fluid flow and solves Stokes equation directly.
+    For further details of this method, read chapter 10 of the thesis: 
+    https://dspace.imsc.res.in/xmlui/handle/123456789/418 
+    or at https://www.imsc.res.in/xmlui/handle/123456789/418
+    """
     def __init__(self, a_, Np_, Lx_, Ly_, Lz_, Nx_, Ny_, Nz_ ):
         self.a = a_
         self.Np = Np_
@@ -46,6 +45,22 @@ cdef class Mima3D:
        
 
     cpdef flowField1s(self, np.ndarray v, double [:] r, double [:] F, double sigma=3, int NN=20):
+        """
+        Compute flow field at field points  due to body forces 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of flow at field points
+            An array of size Nx x Ny x Nz,
+        r: np.array
+            An array of positions
+            An array of size Nx x Ny x Nz,
+        F: np.array
+            An array of body forces
+            An array of size 3*Np,
+        """
         cdef:
             int i, nx, ny, nz , x, y, z, Nx, Ny, Nz, Np, jx, jy, jz
             double arg, facx, facy, facz, kx, ky, kz
@@ -81,6 +96,22 @@ cdef class Mima3D:
 
 
     cpdef flowField2a(self, np.ndarray v, double [:] r, double [:] T, double sigma=3, int NN=20):
+        """
+        Compute flow field at field points  due to body torques 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of flow at field points
+            An array of size Nx x Ny x Nz,
+        r: np.array
+            An array of positions
+            An array of size Nx x Ny x Nz,
+        T: np.array
+            An array of body torques
+            An array of size 3*Np,
+        """
         cdef:
             int i, nx, ny, nz , x, y, z, Nx, Ny, Nz, Np, jx, jy, jz
             double arg, facx, facy, facz, kx, ky, kz
@@ -115,6 +146,22 @@ cdef class Mima3D:
 
 
     cpdef flowField2s(self, np.ndarray v, double [:] r, double [:] S, double sigma=3, int NN=20):
+        """
+        Compute flow field at field points  due to 2s mode of the slip 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of flow at field points
+            An array of size Nx x Ny x Nz,
+        r: np.array
+            An array of positions
+            An array of size Nx x Ny x Nz,
+        V: np.array
+            An array of 2s mode of the slip
+            An array of size 5*Np,
+        """
         cdef:
             int i, nx, ny, nz , x, y, z, Nx, Ny, Nz, Np, jx, jy, jz
             double arg, facx, facy, facz, kx, ky, kz, skx, sky, skz
@@ -152,6 +199,22 @@ cdef class Mima3D:
 
[…]
