+++ docs/requirements.txt
@@ -1,4 +1,0 @@
-cython
-numpy
-scipy
-
