+++ pystokes/periodic.pyx
@@ -9,6 +9,28 @@ cdef double IPI = (2/sqrt(PI))
 @cython.nonecheck(False)
 @cython.wraparound(False)
 cdef class Rbm:
+    """
+    Rigid body motion (RBM) - velocity and angular velocity
+    
+    Methods in this class update velocities or angular velocities 
+    using the inputs of -  arrays of positions, velocity or angular velocity, 
+    along with an array of forces or torques or a slip mode
+
+    The array of velocity or angular velocities is then update by each method. 
+    
+    ...
+
+    ----------
+    radius: float
+        Radius of the particles (a).    
+    particles: int
+        Number of particles (Np)
+    viscosity: float 
+        Viscosity of the fluid (eta)
+    boxSize: float 
+        Length of the box which is reperated periodicly in 3D
+
+   """
     def __init__(self, radius=1, particles=1, viscosity=1.0, boxSize=10):
         self.a   = radius
         self.Np  = particles
@@ -17,6 +39,29 @@ cdef class Rbm:
 
 
     cpdef mobilityTT(self, double [:] v, double [:] r, double [:] F, int Nb=6, int Nm=6):
+        """
+        Compute velocity due to body forces using :math:`v=\mu^{TT}\cdot F` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        F: np.array
+            An array of forces
+            An array of size 3*Np,
+        Nb: int 
+            Number of periodic boxed summed 
+            Default is 10
+        Nm: int 
+            Number of Fourier modes summed
+            Default is 1
+        """
+
         cdef int Np=self.Np, N1=-(Nm/2)+1, N2=(Nm/2)+1, i, j, ii, jj, kk, xx=2*Np, Nbb=2*Nb+1
         cdef double L=self.L,  xi=1.5*sqrt(PI)/L, ixi2 = 1/(xi*xi), mu=1.0/(6*PI*self.eta*self.a), mu1=mu*self.a*0.75, siz=Nb*L
         cdef double a2=self.a*self.a/3, aidr2, k0=2*PI/L, ivol=1/(L*L*L), mt= IPI*xi*self.a*(-3+20*xi*xi*self.a*self.a/3.0), mpp=mu*(1+mt)   # include M^2(r=0)
@@ -82,6 +127,29 @@ cdef class Rbm:
     
     
     cpdef mobilityTR(   self, double [:] v, double [:] r, double [:] T, int Nb=6, int Nm=6):
+        """
+        Compute velocity due to body torque using :math:`v=\mu^{TR}\cdot T` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        T: np.array
+            An array of torques
+            An array of size 3*Np,
+        Nb: int 
+            Number of periodic boxed summed 
+            Default is 6
+        Nm: int 
+            Number of Fourier modes summed
+            Default is 6
+        """
+
         cdef: 
             double L = self.L,  xi = sqrt(PI)/(L)
             double ixi2 = 1/(xi*xi), vx, vy, vz
@@ -136,6 +204,29 @@ cdef class Rbm:
     
     
     cpdef propulsionT2s(self, double [:] v, double [:] r, double [:] S, int Nb=6, int Nm=6):
+        """
+        Compute velocity due to 2s mode of the slip :math:`v=\pi^{T,2s}\cdot S` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
[…]
