+++ pystokes/interface.pyx
@@ -39,6 +39,26 @@ cdef class Rbm:
         self.Mobility = np.zeros( (3*self.Np, 3*self.Np), dtype=np.float64)
 
     cpdef mobilityTT(self, double [:] v, double [:] r, double [:] F, double ll=0):
+        """
+        Compute velocity due to body forces using :math:`v=\mu^{TT}\cdot F` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        F: np.array
+            An array of forces
+            An array of size 3*Np,
+        ll: float 
+            viscosity ratio of the two fluids 
+            Default is zero 
+        """
+
         cdef int i, j, Np=self.Np, xx=2*Np
         cdef double dx, dy, dz, idr, idr3, idr5, Fdotidr, h2, hsq, tempF
         cdef double vx, vy, vz, F1, F2, F3
@@ -128,6 +148,26 @@ cdef class Rbm:
 
 
     cpdef mobilityTR(self, double [:] v, double [:] r, double [:] T, double ll=0):
+        """
+        Compute velocity due to body forces using :math:`v=\mu^{TR}\cdot T` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        T: np.array
+            An array of forces
+            An array of size 3*Np,
+        ll: float 
+            viscosity ratio of the two fluids 
+            Default is zero 
+        """
+
         cdef int Np = self.Np, i, j, xx=2*Np
         cdef double dx, dy, dz, idr, idr3, rlz, Tdotidr, h2,
         cdef double vx, vy, vz, T1, T2, T3
@@ -187,6 +227,23 @@ cdef class Rbm:
 
 
     cpdef propulsionT2s(self, double [:] v, double [:] r, double [:] S, double ll=0):
+        """
+        Compute velocity due to 2s mode of the slip :math:`v=\pi^{T,2s}\cdot S` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        S: np.array
+            An array of 2s mode of the slip
+            An array of size 5*Np,
+        """
+
         cdef int Np=self.Np, i, j, xx=2*Np, xx1=3*Np , xx2=4*Np
         cdef double dx, dy, dz, idr, idr2, idr3, idr5, idr7, aidr2, trS, h2, hsq
         cdef double sxx, syy, szz, sxy, syx, syz, szy, sxz, szx, srr, srx, sry, srz
@@ -314,6 +371,23 @@ cdef class Rbm:
 
 
     cpdef propulsionT3t(self, double [:] v, double [:] r, double [:] D, double ll=0):
+        """
+        Compute velocity due to 3t mode of the slip :math:`v=\pi^{T,3t}\cdot D` 
+        ...
+
+        Parameters
+        ----------
+        v: np.array
+            An array of velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        D: np.array
+            An array of 3t mode of the slip
[…]
