### 2020-06-08 #2299 @harpolea
Hi @rajeshrinet, it's come to my attention that a paper on this library has already been submitted to arXiv (https://arxiv.org/abs/1910.00909). Has this paper been accepted by another journal? If so, then I'm afraid we may not be able to accept this submission as it would involve a significant amount of overlap and duplication with an existing publication.
