### 2020-06-08 #2299 @rajeshrinet
Hi @harpolea,

No https://arxiv.org/abs/1910.00909 has not been accepted (or under consideration) for publication elsewhere. In fact, our submission to JOSS is based on https://arxiv.org/abs/1910.00909 and has been reduced in size to match the word limit. We cite it in the submission as it contains more details. I hope that it okay?
