### 2020-06-08 #2299 @kyleniemeyer
@harpolea @rajeshrinet this seems fine to me. @harpolea is right that it would not be appropriate to have an equivalent software paper under review (or published) at a different journal, but a preprint, whether extended or not, is fine.

I think citing the arXiv eprint in this article as a sort of extended documentation/additional examples is a good idea.
