+++ docs/paper/paper.md
@@ -26,16 +26,14 @@ Hydrodynamic and phoretic interactions between “active particles” in a visco
 
 The principal features that set this method apart are (a) the restriction of independent fluid and phoretic degrees of freedom to the particle boundaries (b) the freedom from grids, both in the bulk of the fluid and on the particle boundaries and (c) the ability to handle, within the same numerical framework, a wide variety of geometries and boundary conditions, including unbounded volumes, volumes bounded by plane walls or interfaces, periodic volumes and, indeed, any geometry-boundary condition combination for which the Green's functions of the governing equations are simply evaluated.
 
-The purpose of this article is to demonstrate the power of the numerical method, as implemented in a Python library, that can be used to simulate experimental phenomena. Our software implementation uses a polylgot programming approach that combines the readability of Python with the speed of Cython and retains the advantages of a high-level, dynamically typed, interpreted language without sacrificing performance. You can take PyStokes for a spin without installation: [Binder demo of PyStokes](https://mybinder.org/v2/gh/rajeshrinet/pystokes/master?filepath=examples). All software is released under the MIT license.
+The purpose of this article is to demonstrate the power of the numerical method, as implemented in a Python library, that can be used to simulate experimental phenomena. Our software implementation uses a polylgot programming approach that combines the readability of Python with the speed of Cython and retains the advantages of a high-level, dynamically typed, interpreted language without sacrificing performance. You can take PyStokes for a spin without installation: [Binder demo of PyStokes](https://mybinder.org/v2/gh/rajeshrinet/pystokes/master?filepath=examples). 
 
 
 # Methods
 
-Our method relies on the reduction of linear elliptic partial differential equations to systems of linear algebraic equations. The four key mathematical steps underpinning it are illustrated in this 
-
+Our method relies on the reduction of linear elliptic partial differential equations to systems of linear algebraic equations. The four key mathematical steps are:
 ![Key mathematical steps underpinning the PyStokes codebase.\label{fig:example}](figure.png)
-
[…]
