+++ pystokes/interface.pyx
@@ -19,14 +19,12 @@ cdef class Rbm:
     Parameters
     ----------
     radius: float
-        Radius of the particles.    
+        Radius of the particles (a)
     particles: int
-        Number of particles 
-    viscosity: viscosity of the fluid 
-    Examples
-    --------
-    An example of the RBM
-
+        Number of particles (Np) 
+    viscosity: float  
+        Viscosity of the fluid (eta)
+    
     """
 
 
@@ -57,6 +55,34 @@ cdef class Rbm:
         ll: float 
             viscosity ratio of the two fluids 
             Default is zero 
+        
+        Examples
+        --------
+        An example of the RBM 
+
+        >>> import pystokes, numpy as np, matplotlib.pyplot as plt
+        >>> # particle radius, self-propulsion speed, number and fluid viscosity
+        >>> b, vs, Np, eta = 1.0, 1.0, 128, 0.1
+        >>> #initialise
+        >>> r = pystokes.utils.initialCondition(Np)  # initial random distribution of positions
+        >>> p = np.zeros(3*Np); p[2*Np:3*Np] = -1    # initial orientation of the colloids
+        >>> 
+        >>> rbm = pystokes.interface.Rbm(radius=b, particles=Np, viscosity=eta)
+        >>> force = pystokes.forceFields.Forces(particles=Np)
+        >>> 
+        >>> def rhs(rp):
+        >>>     # assign fresh values at each time step
+        >>>     r = rp[0:3*Np];   p = rp[3*Np:6*Np]
+        >>>     F, v, o = np.zeros(3*Np), np.zeros(3*Np), np.zeros(3*Np)
+        >>> 
+        >>>     force.lennardJonesWall(F, r, lje=0.01, ljr=5, wlje=1.2, wljr=3.4)
+        >>>     rbm.mobilityTT(v, r, F)
+        >>>     return np.concatenate( (v,o) )
+        >>> 
+        >>> # simulate the resulting system
+        >>> Tf, Npts = 150, 200
+        >>> pystokes.utils.simulate(np.concatenate((r,p)), 
+        >>>    Tf,Npts,rhs,integrator='odeint', filename='crystallization')
         """
 
         cdef int i, j, Np=self.Np, xx=2*Np
@@ -288,7 +314,6 @@ cdef class Rbm:
                     Sljrjy = syx*dx +  syy*dy + syz*dz ;
                     Sljrjz = szx*dx +  szy*dy + szz*dz ;
                     srr = (sxx*dx*dx + syy*dy*dy + szz*dz*dz +  2*sxy*dx*dy)*idr2 ;
-                    #srx = sxy*dx + sxy*dy + szz*dz # old most probably wrong
                     srx = sxx*dx + sxy*dy + sxz*dz
                     sry = syx*dx + syy*dy + syz*dz
                     srz = sxz*dx + syz*dy + szz*dz
@@ -297,28 +322,29 @@ cdef class Rbm:
                     vy += (Sljrly - Sljrjy - trS*dy + 3*srr*dy)*idr3
                     vz += (Sljrlz - Sljrjz - trS*dz + 3*srr*dz)*idr3
 
-                    #vx += -2*(dz*(sxz-3*srz*dx*idr2)+ szz*dx)*idr3;
[…]
