### 2020-06-08 ed0e6fa
update ex and docs
