### 2020-06-08 d3f2bb7
move tests folder in pystokes
