+++ pystokes/tests/testNotebooks.py
@@ -9,12 +9,12 @@ def run_notebook_tests(path, recursive=False):
     Runs Jupyter notebook tests. Exits if they fail.
     """
     basepath = os.path.dirname(__file__)
-    nbpath = os.path.abspath(os.path.join(basepath, "..", path))
+    nbpath = os.path.abspath(os.path.join(basepath, "../..", path))
     '''
     Ignore notebooks which take longer or have deliberate errors, 
     but check they still exists
     '''
-    os.chdir('../examples/')
+    os.chdir('../../examples/')
 
     cwd =os.getcwd()
     ignore_list = [os.path.join(cwd, 'ex3-crystalNucleation.ipynb'),
