+++ docs/paper/paper.md
@@ -22,18 +22,54 @@ bibliography: paper.bib
 
 # Summary
 
-Hydrodynamic and phoretic interactions between “active particles” in a viscous fluid are central to the understanding of their collective dynamics [@ebbens2010pursuit; @zhang2017active]. Under experimentally relevant conditions, the motion of the fluid is governed by the Stokes equation and that of the phoretic field, if one is present, by the Laplace equation. The “activity” appears in these equations as boundary conditions on the particle surfaces that prescribe the slip velocity in the Stokes equation and flux of the phoretic field in the Laplace equation. The slip velocity and the phoretic flux are related by a linear constitutive law that can be derived from a detailed analysis of the boundary layer physics [@anderson1989colloid]. The Stokes and Laplace equations are coupled by this linear constitutive law only at the particle boundaries. The linearity of the governing equations and of the coupling boundary conditions allows for a formally exact solution of the problem of determining the force per unit area on the particle surfaces. This formally exact solution can be approximated to any desired degree of accuracy by a truncated series expansion in a complete basis of functions on the particle boundaries. This, in turn, leads to an efficient and accurate numerical method for computing hydrodynamic and phoretic interactions between active particles [@singh2015many; @singh2018generalized; @singh2019competing].
+Hydrodynamic and phoretic interactions between “active particles” in a viscous fluid are central to the understanding of their collective dynamics [@ebbens2010pursuit; @zhang2017active]. 
+Under experimentally relevant conditions, the motion of the fluid is governed by the Stokes equation and that of the phoretic field, if one is present, by the Laplace equation. 
+The “activity” appears in these equations as boundary conditions on the particle surfaces that prescribe the slip velocity in the Stokes equation and flux of the phoretic field in the Laplace equation. 
+The slip velocity and the phoretic flux are related by a linear constitutive law that can be derived from a detailed analysis of the boundary layer physics [@anderson1989colloid]. 
+The Stokes and Laplace equations are coupled by this linear constitutive law only at the particle boundaries. 
+The linearity of the governing equations and of the coupling boundary conditions allows for a formally exact solution of the problem of determining 
+the force per unit area on the particle surfaces. This formally exact solution can be approximated to any desired degree of accuracy by a 
+truncated series expansion in a complete basis of functions on the particle boundaries. This, in turn, leads to an efficient and accurate numerical 
+method for computing hydrodynamic and phoretic interactions between active particles [@singh2015many; @singh2018generalized; @singh2019competing].
 
-The principal features that set this method apart are (a) the restriction of independent fluid and phoretic degrees of freedom to the particle boundaries (b) the freedom from grids, both in the bulk of the fluid and on the particle boundaries and (c) the ability to handle, within the same numerical framework, a wide variety of geometries and boundary conditions, including unbounded volumes, volumes bounded by plane walls or interfaces, periodic volumes and, indeed, any geometry-boundary condition combination for which the Green's functions of the governing equations are simply evaluated.
+The principal features that set this method apart are (a) the restriction of independent fluid and phoretic degrees of freedom to the particle boundaries 
+(b) the freedom from grids, both in the bulk of the fluid and on the particle boundaries and 
+(c) the ability to handle, within the same numerical framework, a wide variety of geometries and boundary conditions, 
+including unbounded volumes, volumes bounded by plane walls or interfaces, periodic volumes and, indeed, 
+any geometry-boundary condition combination for which the Green's functions of the governing equations are simply evaluated.
 
[…]
