### 2020-06-08 ce4e718
add description in examples
