+++ examples/plotter.py
@@ -1,487 +1,0 @@
-import  numpy as np
-import matplotlib.pyplot as plt
-from matplotlib import cm, colors
-from mpl_toolkits.mplot3d import Axes3D
-from matplotlib import rc
-rc('text', usetex=True)
-
-
-
-def gridXY(dim, L, Ng):
-    """
-    returns the grid in XY direction centered around zero
-    """
-    Nt = Ng*Ng
-    rr, vv = np.zeros(dim*Nt), np.zeros(dim*Nt)
-    X, Y = np.meshgrid(np.linspace(-L, L, Ng), np.linspace(-L, L, Ng))
-    rr[0:2*Nt] = np.concatenate((X.reshape(Ng*Ng), Y.reshape(Ng*Ng)))
-    return rr, vv
-
-
-def gridYZ(dim, L, Ng):
-    """
-    returns the grid in YZ direction centered around zero
-    """
-    Nt = Ng*Ng
-    rr, vv = np.zeros(dim*Nt), np.zeros(dim*Nt)
-    X, Y = np.meshgrid(np.linspace(-L, L, Ng), np.linspace(-0, L, Ng))
-    rr[Nt:3*Nt] = np.concatenate((X.reshape(Ng*Ng), Y.reshape(Ng*Ng)))
-    return rr, vv
-
-
-
-def simulate(rp0, Tf, Npts, rhs, integrator='odeint', filename='this'):
-    from scipy.io import savemat; from scipy.integrate import odeint
-
-    time_points=np.linspace(0, Tf, Npts+1);
-    def rhs0(rp, t): 
-        return rhs(rp)
-    
-    if integrator=='odeint':
-        u = odeint(rhs0, rp0, time_points, mxstep=5000000)
-        savemat(filename, {'X':u, 't':time_points})
-
-    elif integrator=='odespy-vode':
-        import odespy
-        solver = odespy.Vode(rhs0, method = 'bdf', atol=1E-7, rtol=1E-6, order=5, nsteps=10**6)
-        solver.set_initial_condition(rp0)
-        u, t = solver.solve(time_points) 
-        savemat(filename, {'X':u, 't':time_points})
-
-    else:
-        print("Error: Integration failed! \n Please set integrator='odeint' to use the scipy odeint (Deafult). \n Use integrator='odespy-vode' to use vode from odespy (github.com/rajeshrinet/odespy). \n Alternatively, write your own integrator to evolve the system in time \n")
-    return
-
-
-def initialCondition(Np, h0=3.1):
-    '''
-    Assigns initial condition. 
-    To DO: Need to include options to assign possibilites
-    '''
-    rp0 = np.zeros(3*Np) 
-    radius = np.sqrt(np.arange(Np)/float(Np))
-    golden_angle = np.pi * (3 - np.sqrt(5))
-    theta = golden_angle * np.arange(Np)
-    
-    points = np.zeros((Np, 2)) 
-    points[:,0] = np.cos(theta)
-    points[:,1] = np.sin(theta)
-    points *= radius.reshape((Np, 1)) 
-    points = points*2.1*np.sqrt(Np)
-    
-    points[:,0] += 2*(2*np.random.random(Np)-np.ones(Np)) 
-    points[:,1] += 2*(2*np.random.random(Np)-np.ones(Np)) 
-    
-    
-    rp0[0:Np]          = 1.5*points[:, 0]
-    rp0[Np:2*Np]   = 1.5*points[:, 1]
-    rp0[2*Np:3*Np] = h0
-    ##rp0[0:3*Np] = ar*(2*np.random.random(3*Np)-np.ones(3*Np)) + rp0[0:3*Np]
-    #
[…]
