### 2020-06-09 #2318 @whedon
[REVIEW]: PyStokes: Stokesian hydrodynamics in Python
**Submitting author:** @rajeshrinet (<a href="http://orcid.org/0000-0003-0266-9691">Rajesh Singh</a>)
**Repository:** <a href="https://github.com/rajeshrinet/pystokes" target ="_blank">https://github.com/rajeshrinet/pystokes</a>
**Version:** v2.2.0
