### 2020-06-22 717871b
update description of optional odespy #3
