### 2020-06-22 5218798
add subset of example folder for binder demo
