+++ binder/README.md
@@ -1,0 +1,1 @@
+The full set of examples is at: https://github.com/rajeshrinet/pystokes/tree/master/examples
