+++ examples/others/benchmarks/README.md
@@ -1,9 +1,0 @@
-## PyStokes: Benchmark of PyStokes
-
-
-The library scales linearly with the number of cores while an *O(M^2)* computational effort for *M* unknowns when matrix-vector products are computed directly as shown below. 
-
-![Alt text](benchmark.jpg "Benchmarks for calculation of the propulsion tensor corresponding to potential dipole on a 16-core machine using the PyStokes library")
-
-
-The above benchmark of PyStokes library is done on a 16-core machine for the evaluation of propulsion tensor corresponding to potential dipole.
