+++ examples/others/forceFields/README.md
@@ -1,0 +1,1 @@
+Examples of force fields which can be implemented in PyStokes. 
