### 2020-06-22 rajeshrinet/pystokes#6 @khinsen
Provide instructions for contributing
JOSS requires "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". 2) and 3) are discussed in the README, but I didn't find guidelines for potential contributors.
