### 2020-06-22 2b76ce0
add contributing #6
