+++ CONTRIBUTING.md
@@ -1,0 +1,4 @@
+CONTRIBUTING
+============
+
+When contributing code to PyStokes, you agree to release it under the project's license. Please first discuss the change you wish to make via issue, email, or any other method with the owners. 
