### 2020-06-22 rajeshrinet/pystokes#6 @rajeshrinet
@khinsen I have now added a new file: https://github.com/rajeshrinet/pystokes/blob/master/CONTRIBUTING.md with a brief guidelines for potential contributors.
