### 2020-06-22 #2318 @khinsen
@harpolea I did a first pass looking at the code and the paper. It was a lot of fun: this is a very nice implementation of hydrodynamic and phoretic interactions which invites users to play with it, in particular by including visualization tools, and that is just great for doing research. The next time I do work in this domain, I'll happily abandon [my own 25-year-old Fortran library](https://github.com/khinsen/HYDROLIB) and use PyStokes instead!

I have made a few suggestions for improvements via issues/pull requests on the code repository, which the authors have handled immediately. Interactive reviewing is a fun experience :-)

There are a few open points in the review checklist as of today:

1. There are no tests, only examples. This being number-crunching code, it is very difficult to write tests that are meaningful and predictable, given how little control programmers have over the details of float arithmetic (shameless self-advertising: [The Approximation Tower in Computational Science: Why Testing Scientific Software Is Difficult](https://hal.archives-ouvertes.fr/hal-01171382)). Does JOSS have an editorial policy concerning tests for number-crunching software? Or at least recommendations for authors? I didn't find anything.

2. Community guidelines: There are instructions for reporting issues and for seeking support, but not for contributing to the software. See https://github.com/rajeshrinet/pystokes/issues/6.

3. High-level summary in the paper: The current summary is clear enough for potential users of the software, but maybe not "for a diverse, non-specialist audience". But this should best be judged by someone unfamiliar with the field.

4. State of the field: the paper doesn't provide comparisons with other commonly-used packages, but then there are no competing commonly used packages, as far as I can tell. But the authors should definitely say something about this, even if it's only a confirmation of my suspicion.
