### 2020-06-22 #2318 @khinsen
Contribution guidelines were just added!
