### 2020-06-22 #2318 @harpolea
@khinsen Thank you for your review!

As you probably read, the JOSS reviewer criteria for testing can be found [here](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). While not required, it is strongly encouraged that some kind of automated testing suite is set up to test the functionality of the code.

As you said, testing codes such as these can be tricky. I am not that familiar with this particular branch of CFD, however I imagine there may be some simple, standard test problems with (semi-)analytic solutions? E.g. simple advection tests, tests to demonstrate energy conservation, etc

Regression tests could also be another option. In the codes I work on, we save the output from a selection of problems as benchmarks, then test against these whenever changes are made to the code to ensure that they do not change the output.
