### 2020-06-22 #2318 @rajeshrinet
@khinsen and @harpolea there are some tests, which run every time someone pushes to the repo (it runs four tests and checks all the notebooks). the folder contains tests is somewhat hidden: pystokes/tests: https://github.com/rajeshrinet/pystokes/tree/master/pystokes/tests

It does the test of [Stoke's law](https://en.wikipedia.org/wiki/Stokes%27_law) for rotation and translation in an unbounded domain and near a plane wall.
