### 2020-06-22 3b5ef34
add test to compute mobility of a sedimenting lattice and compare it with Zick and Homsy 1982
