### 2020-06-22 #2318 @rajeshrinet
I have now added one more test: compare mobility of a sedimenting lattice with known results in literature (Zick and Homsy, 1982) and moved `tests`  at start of folder tree.
