### 2020-06-23 #2318 @khinsen
@rajeshrinet I had indeed overlooked the tests! Thanks for making them more visible. And yes, the old Zick/Homsy paper makes for a nice test based on well-known results. Could you add instructions to the README for how to run the tests? That would also ensure their visibility.
