### 2020-06-23 c7fa284
add testing commands in readme openjournals/joss-reviews/#2318
