+++ tests/installTests.py
@@ -1,13 +1,12 @@
-#!python
-"""Unittesting for the pystokes module. Run as python -m unittest pystokes.test."""
+"""
+Unittesting for pystokes
+"""
 import sys
 import pystokes
 import unittest
 import inspect
 import numpy as np
 import scipy as sp
-
-
 
 
 class UnboundedTest(unittest.TestCase):
@@ -28,8 +27,6 @@ class UnboundedTest(unittest.TestCase):
         diff = V1[2] - V2[2] 
         self.assertTrue((np.asarray(diff) < 0.001).all(),
                        msg=f"Stokes law for translation is not satisfied")
-
- 
 
 
     def test_rotation(self):
@@ -54,6 +51,7 @@ class UnboundedTest(unittest.TestCase):
 
 class WallBoundedTest(unittest.TestCase):
     
+
     def test_parallelTranslation(self):
         r = np.array([0,0,1.])
         F = np.array([0,1,0.])
@@ -95,6 +93,8 @@ class WallBoundedTest(unittest.TestCase):
 
 
 class PeriodicTest(unittest.TestCase):
+
+
     def test_effectiveMobility(self):
         a, eta, Np = 1.0, 1.0/6, 1
         v = np.zeros(3*Np)
@@ -108,7 +108,7 @@ class PeriodicTest(unittest.TestCase):
 
         mu=1.0/(6*np.pi*eta*a)
         diff = -v[2]/mu - 0.498
-        self.assertTrue((np.asarray(diff) < 0.01).all(),
+        self.assertTrue((np.asarray(diff) < 0.002).all(),
                        msg=f"Effective mobility does not match Zick & Homsy (1982)")
 
 
