+++ setup.py
@@ -1,6 +1,6 @@
 import numpy, os, sys, os.path, tempfile, subprocess, shutil
 import os, sys
-#from setuptools import setup, Extension
+from setuptools import setup, Extension
 from distutils.core import setup
 from Cython.Build import cythonize
 from distutils.extension import Extension
@@ -54,7 +54,7 @@ else:
 
 setup(
     name='pystokes',
-    version='2.1.1',
+    version='2.1.2',
     url='https://github.com/rajeshrinet/pystokes',
     author = 'The PyStokes team',
     author_email = 'PyStokes@googlegroups.com',
@@ -72,10 +72,7 @@ setup(
         ),
     libraries=[],
     packages=['pystokes'],
-    #install_requires=['cython','numpy','scipy'],
-    #extras_require={
-    #        'plotting': ['matplotlib'],
-    #        'notebook': ['jupyter', 'nbconvert']},
+    install_requires=['cython','numpy','scipy','matplotlib','jupyter','nbconvert'],
     package_data={'pystokes': ['*.pxd']}
 )
 
