+++ binder/environment.yml
@@ -3,13 +3,6 @@ channels:
   - conda-forge
 dependencies:
   - python
-  - numpy
-  - matplotlib
-  - cython
-  - scipy 
-  - jupyter
-  - ipython
-  - nbconvert
   - pip
   - pip:
     - pystokes
