### 2020-06-23 rajeshrinet/pystokes#10 @fcooper8472
Reporting coverage
I'm not sure how much effort would be required for this, but it would be great to automatically generate a code coverage report so you could see what proportion of the code is currently covered by tests.

If this was a pure Python library, it would be very straightforward using [coverage](https://coverage.readthedocs.io/en/coverage-5.1/) to generate the coverage info and [codecov](https://codecov.io/) to produce reports.

However, because this is a Cython library, I am not sure what would be required to get coverage reporting up-and-running, so for now this is only open as a suggestion.
