### 2020-06-23 rajeshrinet/pystokes#10 @rajeshrinet
@fcooper8472 thanks for the suggestion. I will read through their documentation and try to get it done for PyStokes.

I must admit that the current testing is very minimal.
