+++ requirements.txt
@@ -2,3 +2,6 @@ numpy
 matplotlib
 cython
 scipy 
+jupyter
+ipython
+nbconvert
