### 2020-06-24 #2318 @khinsen
@rajeshrinet @harpolea Looks good (and the tests all pass on my machine)! I am happy with the software in its current state. I have suggestions for a revision of the paper (see above), but you may prefer to wait for @fcooper8472's feedback before attacking the paper.
