### 2020-06-24 38d25db
added more examples
