+++ examples/others/mima/README.md
@@ -2,4 +2,4 @@
 Mollified irreducible multipole approach (MIMA) resolves the fluid flow and solves Stokes equation directly.  
 
 
-The .ipynb files show the usage. 
+The .ipynb files illustrate the usage. 
