+++ examples/others/flow_confinement/README.md
@@ -1,1 +1,0 @@
-This folder contains example on how to use PyStokes to study the role of confinement near a plane wall. 
