### 2020-06-24 #2318 @fcooper8472
I think the software looks very good.

I found a few problems with installation, but @rajeshrinet addressed those very quickly - it's great for this to be happening in real time!

In terms of the formal checklist, I have the same two checkboxes un-checked as @khinsen:

- **Summary**: I have some experience in this area (I once coded up a spectral method for solving Navier-Stokes on a periodic domain), but I am not a domain expert and do not understand the whole of the first paragraph.  Even after a bit of time on Google I'm still not certain I know what the phoretic field is, for instance.  I think a "diverse non-specialist audience" would benefit from a paragraph up-front with a higher level overview of what the software is designed to do.

- **State of the field**: I cannot tell from reading the paper whether there is other software available that does a similar task.  Either way, I think this needs stating in the paper.

Other than that, a couple of suggestions:

- The README (until very recently) stated that PyStokes works with Python 2.6+ or Python 3.4+: I've not been able to verify this, but it would be very useful for users to know which versions it works with.  It would be very easy to add a GitHub workflow to test on multiple Python version.  @rajeshrinet would you be interested in a PR that implements that?

- Testing (outside of checking that examples run - which is great!) is fairly thin.  I would encourage coverage checking (suggested [here](https://github.com/rajeshrinet/pystokes/issues/10) to assess what proportion of code is covered by tests, but on top of that it would be great to see a unit testing framework in use.  I understand that some of the number crunching can be very hard to test effectively: but even boring things like the utils would benefit from some testing.
