### 2020-06-25 bc8863c
mention support for py 2 and 3
