+++ setup.py
@@ -7,17 +7,17 @@ Cython.Compiler.Options.annotate=True
 
 
 def checkOpenmpSupport():
-    """ Adapted from https://stackoverflow.com/questions/16549893/programatically-testing-for-openmp-support-from-a-python-setup-script
-    """ 
+    ''' Adapted from https://stackoverflow.com/questions/16549893/programatically-testing-for-openmp-support-from-a-python-setup-script
+    ''' 
     ompTest = \
-    r"""
+    r'''
     #include <omp.h>
     #include <stdio.h>
     int main() {
     #pragma omp parallel
-    printf("Thread %d, Total number of threads %d\n", omp_get_thread_num(), omp_get_num_threads());
+    printf('Thread %d, Total number of threads %d\n', omp_get_thread_num(), omp_get_num_threads());
     }
-    """
+    '''
     tmpdir = tempfile.mkdtemp()
     curdir = os.getcwd()
     os.chdir(tmpdir)
@@ -30,7 +30,7 @@ def checkOpenmpSupport():
             result = subprocess.call(['cc', '-fopenmp', filename],
                                      stdout=fnull, stderr=fnull)
     except:
-        print("Failed to test for OpenMP support. Assuming unavailability");
+        print('Failed to test for OpenMP support. Assuming unavailability');
         result = -1;
     
     os.chdir(curdir)
@@ -51,28 +51,43 @@ else:
 with open('requirements.txt', 'r') as rm:
     reqs = [l.strip() for l in rm]
 
+with open("README.md", "r") as fh:
+    long_description = fh.read()
+
 
 setup(
     name='pystokes',
-    version='2.1.5',
+    version='2.1.6',
     url='https://github.com/rajeshrinet/pystokes',
     author = 'The PyStokes team',
     author_email = 'PyStokes@googlegroups.com',
     license='MIT',
     description='Stokesian hydrodynamics in Python',
-    long_description='Pystokes is a library for computing \
-    hydrodynamic and phoretic interactions of active particles in Python',
+    long_description=long_description,
+    long_description_content_type='text/markdown',
     platforms='tested on LINUX',
-    ext_modules=cythonize([ Extension("pystokes/*", ["pystokes/*.pyx"],
+    ext_modules=cythonize([ Extension('pystokes/*', ['pystokes/*.pyx'],
         include_dirs=[numpy.get_include()],
         extra_compile_args=ompArgs,
         extra_link_args=ompArgs,
         )],
-        compiler_directives={'language_level' : "3"},
+        compiler_directives={'language_level': sys.version_info[0]},
         ),
     libraries=[],
     packages=['pystokes'],
-    install_requires=reqs,
-    package_data={'pystokes': ['*.pxd']}
+    package_data={'pystokes': ['*.pxd']},
+    include_package_data=True,
+    classifiers=[
+        'License :: OSI Approved :: MIT License',
+        'Programming Language :: Python',
+        'Programming Language :: Python :: 2',
+        'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.4',
+        'Programming Language :: Python :: 3.5',
+        'Programming Language :: Python :: 3.6',
+        'Programming Language :: Python :: 3.7',
+        ],
+    install_requires=reqs
 )
 
