+++ .github/workflows/build-test.yml
@@ -11,6 +11,9 @@ jobs:
   build-and-test:
     name: Install and test
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [2.7, 3.5, 3.6, 3.7]
 
     steps:
       - uses: actions/checkout@v2
