### 2020-06-25 #2318 @rajeshrinet
@fcooper8472 and @khinsen many thanks for your comments above and for several suggestions on the repo which have already helped improved the codebase. I am now working on the paper and would update it in the light of your comments.  I will post a comment here when done.

@fcooper8472, indeed PyStokes was first written in Python 2 and was later modified to support Python 3. Thus, it supports both Python 2 and 3. I have now added GitHub workflow which checks both Python 2 and 3. See the result at: https://github.com/rajeshrinet/pystokes/runs/806434265. I now mention this in the README as well. I am now working on getting the coverage bit up and running.
