+++ .github/workflows/build-test.yml
@@ -18,18 +18,12 @@ jobs:
     steps:
       - uses: actions/checkout@v2
     
-      - name: Setup Miniconda
-        uses: goanpeca/setup-miniconda@v1.2.0
-        with:
-         activate-environment: pystokes
-         environment-file: environment.yml
-         # auto-activate-base: true
-      - name: make-pystokes
+      - name: install-pystokes
         shell: bash -l {0}
         run: |
-          conda activate pystokes
+          python setup.py install
           make
       - name: test-pystokes
         shell: bash -l {0}
         run: |
-          make test
+          cd tests && python shortTests.py
