+++ .github/workflows/build-test.yml
@@ -13,17 +13,22 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [2.7, 3.6]
-
+        python-version: [2.7, 3.7]
     steps:
       - uses: actions/checkout@v2
     
-      - name: install-pystokes
+      - name: Setup Miniconda
+        uses: goanpeca/setup-miniconda@v1.2.0
+        with:
+         activate-environment: pystokes
+         environment-file: environment.yml
+         python-version: ${{ matrix.python-version }}
+      - name: make-pystokes
         shell: bash -l {0}
         run: |
-          pip install -r requirements.txt && python setup.py install
+          conda activate pystokes
           make
       - name: test-pystokes
         shell: bash -l {0}
         run: |
-          cd tests && python shortTests.py
+          make test
