### 2020-06-25 b2eba82
minor fix to make sure test runs in 2.7
