+++ pystokes/utils.pyx
@@ -5,14 +5,10 @@ and plotting flow and phoretic fields
 
 import  numpy as np
 cimport numpy as np
-import matplotlib.pyplot as plt
 cimport cython
 from libc.math cimport sqrt, pow, log
 from cython.parallel import prange
 cdef double PI = 3.1415926535
-from matplotlib import rc
-from matplotlib import cm, colors
-from mpl_toolkits.mplot3d import Axes3D
 
 DTYPE   = np.float
 DTYP1   = np.int32
@@ -130,6 +126,8 @@ cpdef irreducibleTensors(l, p, Y0=1):
     return YY
 
 
+
+
 def simulate(rp0, Tf, Npts, rhs, integrator='odeint', filename='this.mat'):
     """
     Simulates using choice of integrator
@@ -199,6 +197,8 @@ def simulate(rp0, Tf, Npts, rhs, integrator='odeint', filename='this.mat'):
     return
 
 
+
+
 def initialCondition(Np, h0=3.1):
     '''
     Assigns initial condition. 
@@ -228,6 +228,8 @@ def initialCondition(Np, h0=3.1):
     #rp0[4*Np:5*Np] = np.zeros(Np)
     #rp0[5*Np:6*Np] = -np.ones(Np)
     return rp0 
+
+
 
 
 def plotLogo():
@@ -280,12 +282,15 @@ def plotLogo():
     plt.text(r[Np+3]-.46, r[2*Np+3]-.54, 'es', fontsize=111);
 
 
+
+
 def plotStreamlinesXY(vv, rr, r, density=0.82, arrowSize=1.2, mask=0.6, ms=36, offset=1e-6, title='None'):
     """
     Plots streamlines in XY plane given the position and velocity; 
     vv: one dimensional arrays of velocity
     rr: one dimensional arrays of positions where velocity is computed
     """
+    import matplotlib.pyplot as plt
     Np, Nt = int(np.size(r)/3), int(np.size(rr)/3);  Ng=int(np.sqrt(Nt))
     xx, yy = rr[0:Nt].reshape(Ng, Ng), rr[Nt:2*Nt].reshape(Ng, Ng)
     vx, vy = vv[0:Nt].reshape(Ng, Ng), vv[Nt:2*Nt].reshape(Ng, Ng)
@@ -322,12 +327,15 @@ def plotStreamlinesXY(vv, rr, r, density=0.82, arrowSize=1.2, mask=0.6, ms=36, o
         plt.title(title, fontsize=26);
 
 
+
+
 def plotStreamlinesYZ(vv, rr, r, density=0.795, arrowSize=1.2, mask=0.6, ms=36, offset=1e-6, title='None'):
     """
     Plots streamlines in YZ plane given the position and velocity; 
     vv: one dimensional arrays of velocity
     rr: one dimensional arrays of positions where velocity is computed
     """
+    import matplotlib.pyplot as plt
     Np, Nt = int(np.size(r)/3), int(np.size(rr)/3);  Ng=int(np.sqrt(Nt))
     yy, zz = rr[Nt:2*Nt].reshape(Ng, Ng), rr[2*Nt:3*Nt].reshape(Ng, Ng)
     vy, vz = vv[Nt:2*Nt].reshape(Ng, Ng), vv[2*Nt:3*Nt].reshape(Ng, Ng)
@@ -363,12 +371,15 @@ def plotStreamlinesYZ(vv, rr, r, density=0.795, arrowSize=1.2, mask=0.6, ms=36, 
         plt.title(title, fontsize=26);
 
 
+
+
 def plotStreamlinesYZsurf(vv, rr, r, density=0.8, arrowSize=1.2, mask=0.6, ms=36, offset=1e-6, title='None'):
     """
[…]
