### 2020-06-25 #2318 @fcooper8472
@rajeshrinet this is great!

Just to be clear I do not think the general suggestions I have made about PyStokes are strictly required for publication in JOSS, though it is great that you're addressing these.
