+++ setup.py
@@ -54,7 +54,7 @@ with open("README.md", "r") as fh:
 
 setup(
     name='pystokes',
-    version='2.1.7',
+    version='2.1.8',
     url='https://github.com/rajeshrinet/pystokes',
     author = 'The PyStokes team',
     author_email = 'PyStokes@googlegroups.com',
