+++ .github/workflows/build-test.yml
@@ -1,10 +1,6 @@
 name: Installation
 
-
-# Controls when the action will run. Triggers the workflow on push or pull request
-# events but only for the master branch
 on: push
-
 
 jobs:
 
@@ -32,3 +28,10 @@ jobs:
         shell: bash -l {0}
         run: |
           make test
+
+      - name: Upload coverage reports to Codecov
+      run: |
+        codecov
+      env: 
+        super_secret: ${{ secrets.CODECOV_TOKEN }}
+
