+++ .github/workflows/build-test.yml
@@ -29,9 +29,11 @@ jobs:
         run: |
           make test
 
-      - name: Upload coverage reports to Codecov
-      run: |
-        codecov
-      env: 
-        super_secret: ${{ secrets.CODECOV_TOKEN }}
+      - name: Upload coverage to codecov
+      uses: codecov/codecov-action@v1 
+      with:
+        flags: unittests
+        env_vars: OS,PYTHON
+        name: codecov-umbrella
+        fail_ci_if_error: true
 
