+++ .github/workflows/coverage.yml
@@ -1,0 +1,12 @@
+name: Example workflow for Codecov
+on: [push, pull_request]
+jobs:
+  run:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@master
+    - name: Upload coverage to Codecov  
+      uses: ./
+      with:
+        flags: unittest
+        name: codecov-
