+++ .travis.yml
@@ -1,0 +1,35 @@
+language: python
+# cache: pip
+notifications:
+  email: false
+dist: xenial
+sudo: false
+python:
+  - 2.7
+  - 3.7
+matrix:
+  include:
+    - python: 2.7
+      dist: trusty
+      virtualenv:
+        system_site_packages: true
+      addons:
+        apt:
+          packages:
+            - python-requests
+            - python-coverage
+            - python-mock
+    - python: 3.7
+      dist: xenial
+      script: make format
+
+install:
+  - pip install -r requirements.txt
+  - pip install --upgrade setuptools wheel
+  - pip install coverage
+  - python setup.py install
+script:
+  - make 
+  - make test
+after_success:
+  - codecov
