+++ .github/workflows/notebook-test.yml
@@ -1,4 +1,4 @@
-name: Notebooks
+name: notebooks
 
 on: push
 
