+++ .github/workflows/test-coverage.yml
@@ -26,8 +26,6 @@ jobs:
           # make nbtest
       - name: Generate coverage report
         run: |
-          pip install pytest 
-          pip install pytest-cov
           pytest tests/shortTests.py --cov=./ --cov-report=xml
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
