+++ .github/workflows/notebook-test.yml
@@ -1,6 +1,6 @@
 name: notebooks
 
-on: push
+on: [push, pull_request]
 
 
 jobs:
