### 2020-06-25 rajeshrinet/pystokes#10 @rajeshrinet
@fcooper8472 I have now added [![codecov](https://codecov.io/gh/rajeshrinet/pystokes/branch/master/graph/badge.svg)](https://codecov.io/gh/rajeshrinet/pystokes). Will add more testing!
