### 2020-06-25 rajeshrinet/pystokes#10 @rajeshrinet
in continuation, as you have noted it only tests .py files. I will see if it can be extended to check .pyx files as well.
