+++ .github/workflows/test-coverage.yml
@@ -23,7 +23,7 @@ jobs:
         run: |
           conda activate pystokes
           make
-          pytest tests/shortTests.py --cov=./ --cov-report=xml 
+          pytest --cov=./ --cov-report=xml
       - uses: codecov/codecov-action@v1.0.2
         with:
           token: ${{secrets.CODECOV_TOKEN}} #required
