### 2020-06-26 abd1216
make statement for cov
