+++ .github/workflows/test-coverage.yml
@@ -21,7 +21,8 @@ jobs:
       - name: make-pystokes-cov
         shell: bash -l {0}
         run: |
-          conda activate pystokes
+          conda activate pystokes 
+          make
           make cycov 
       - uses: codecov/codecov-action@v1.0.2
         with:
