+++ setup.py
@@ -55,7 +55,7 @@ with open("README.md", "r") as fh:
 
 setup(
     name='pystokes',
-    version='2.1.8',
+    version='2.1.10',
     url='https://github.com/rajeshrinet/pystokes',
     author = 'The PyStokes team',
     author_email = 'PyStokes@googlegroups.com',
