### 2020-06-27 rajeshrinet/pystokes#10 @rajeshrinet
@fcooper8472 as you pointed at the start of this issue, coverage in Cython codebase is a bit tricky. I did an example here: https://github.com/rajeshrinet/codecov-cython.

I do not like ways in which I had to change the setup.py and force install the codebase for line coverage. It also leads to performance loss as explained here: http://blog.behnel.de/posts/coverage-analysis-for-cython-modules.html

Thus, I have decided to not implement it for PyStokes as I would like to squeeze every bit of performance possible. But, I would add a more complete unit tests gradually.
