### 2020-06-28 288e3ef
add cite to hydrolib and libstokes
