+++ docs/paper/paper.bib
@@ -244,3 +244,31 @@
 	doi={10.1073/pnas.1718807115},
 	year={2018},
 }
+
+@article{hinsen1995,
+    title={HYDROLIB: a library for the evaluation of hydrodynamic interactions in colloidal suspensions},
+    author={Hinsen, K.},
+    journal={Comp. Phys. Comm.},
+    volume={88},
+    number={2},
+    pages={327--340},
+    year={1995},
+    publisher={Elsevier}
+}
+
+@misc{libstokes,
+    author={Ichiki, K.},
+    title={Py{R}itz},
+    howpublished={\url{https://github.com/kichiki/libstokes}},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    year={2014},
+} 
+
+@book{kim2005,
+    title={Microhydrodynamics: Principles and Selected Applications},
+    author={Kim, S. and Karrila, S. J.},
+    publisher={Butterworth-Heinemann, Boston},
+    doi={10.1002/cite.330641004},
+    year={1991}
+}
