### 2020-06-28 32c98d3
Update paper.md
