+++ docs/paper/paper.md
@@ -22,7 +22,9 @@ bibliography: paper.bib
 
 # Summary
 
-Hydrodynamic and phoretic interactions between “active particles” in a viscous fluid are central to the understanding of their collective dynamics [@ebbens2010pursuit; @zhang2017active]. 
+The PyStokes library allows to study active particles in a Stokesian fluid [@kim2005]. Active particles are distinguished by their ability to produce flow, and thus motion, in the absence of external forces or torques on them. Examples of active particles include auto-phoretic synthetic particles [@ebbens2010pursuit], and microorganisms [@brennen1977]. An auto-phoretic particle self-propels due to non-equilibrium processes, such as chemical reactions on its surface [@anderson1989colloid]. Examples of phoresis (phoretic field) are electrophoresis (electric field), diffusiophoresis (concentration of chemical field), thermophoresis (temperature field) etc [@anderson1989colloid]. A particle is "active" if it creates the phoretic field itself (for example, by a built-in chemical asymmetry on its surface), and the phenomena is called self-phoresis [@ebbens2010pursuit; @zhang2017active]. 
+
+Hydrodynamic and phoretic interactions between active particles in a viscous fluid are central to the understanding of their collective dynamics. 
 Under experimentally relevant conditions, the motion of the fluid is governed by the Stokes equation and that of the phoretic field, if one is present, by the Laplace equation. 
 The “activity” appears in these equations as boundary conditions on the particle surfaces that prescribe the slip velocity in the Stokes equation and flux of the phoretic field in the Laplace equation. 
 The slip velocity and the phoretic flux are related by a linear constitutive law that can be derived from a detailed analysis of the boundary layer physics [@anderson1989colloid]. 
@@ -32,20 +34,23 @@ the force per unit area on the particle surfaces. This formally exact solution c
 truncated series expansion in a complete basis of functions on the particle boundaries. This, in turn, leads to an efficient and accurate numerical 
 method for computing hydrodynamic and phoretic interactions between active particles [@singh2015many; @singh2018generalized; @singh2019competing].
 
-The principal features that set this method apart are (a) the restriction of independent fluid and phoretic degrees of freedom to the particle boundaries 
-(b) the freedom from grids, both in the bulk of the fluid and on the particle boundaries and 
-(c) the ability to handle, within the same numerical framework, a wide variety of geometries and boundary conditions, 
-including unbounded volumes, volumes bounded by plane walls or interfaces, periodic volumes and, indeed, 
-any geometry-boundary condition combination for which the Green's functions of the governing equations are simply evaluated.
+
+![Library structure: input and output of the library and equations that determine the hydrodynamic and phoretic interactions between active particles in a three-dimensional domain $V$, how these “blocks” are coupled at the particle boundaries $S_{i}$. Particle indices are $i=1,\ldots,N$ and harmonic indices are $l=1,2,\ldots$ and $\sigma=s,a,t$. {fig:figS}](FigSchema.png)
+
[…]
