+++ docs/paper/paper.bib
@@ -272,3 +272,15 @@
     doi={10.1002/cite.330641004},
     year={1991}
 }
+
+@article{brennen1977,
+    author = "Brennen, C. and Winet, H.",
+    doi = "10.1146/annurev.fl.09.010177.002011",
+    journal = "Annu. Rev. Fluid Mech.",
+    number = "1",
+    pages = "339--398",
+    publisher = "Annual Reviews 4139 El Camino Way, PO Box 10139, Palo Alto, CA 94303-0139, USA",
+    title = "{Fluid mechanics of propulsion by cilia and flagella}",
+    volume = "9",
+    year = "1977"
+}
