+++ README.md
@@ -24,9 +24,10 @@
 
 [PyStokes](https://github.com/rajeshrinet/pystokes) is a numerical library for phoresis and Stokesian hydrodynamics in Python. It uses a grid-free method, combining the integral representation of Laplace and Stokes equations, spectral expansion, and Galerkin discretization, to compute phoretic and hydrodynamic interactions between spheres with slip boundary conditions on their surfaces. The library also computes suspension scale quantities, such as rheological response, energy dissipation and fluid flow. The computational cost is quadratic in the number of particles and upto 1e5 particles have been accommodated on multicore computers. The library has been used to model suspensions of **microorganisms**,  **synthetic autophoretic particles** and **self-propelling droplets**. 
 
-A detailed description of above keywords is available in the [Wiki](https://github.com/rajeshrinet/pystokes/wiki). Please read the [Wiki](https://github.com/rajeshrinet/pystokes/wiki) before you use PyStokes for your research.  
+A detailed description of above keywords is available in the [Wiki](https://github.com/rajeshrinet/pystokes/wiki). Please read the [Wiki](https://github.com/rajeshrinet/pystokes/wiki) before you use PyStokes for your research. Included below are some examples from [PyStokes Gallery](https://github.com/rajeshrinet/pystokes/wiki/Gallery)
 
-![Crystallization of active colloids](https://raw.githubusercontent.com/rajeshrinet/pystokes/master/examples/crystallite.gif)
+![Image](https://raw.githubusercontent.com/rajeshrinet/pystokes-examples/master/gallery/2_volvox.gif) 
+![Crystallization of active colloids](https://raw.githubusercontent.com/rajeshrinet/pystokes/master/examples/crystallite.gif) 
 
 
 ## News
