+++ docs/paper/paper.md
@@ -22,11 +22,14 @@ bibliography: paper.bib
 
 # Summary
 
-The PyStokes library allows to study active particles in a Stokesian fluid [@kim2005]. Active particles are distinguished by their ability to produce flow, and thus motion, in the absence of external forces or torques on them. Examples of active particles include auto-phoretic synthetic particles [@ebbens2010pursuit], and microorganisms [@brennen1977]. An auto-phoretic particle self-propels due to non-equilibrium processes, such as chemical reactions on its surface [@anderson1989colloid]. Examples of phoresis (phoretic field) are electrophoresis (electric field), diffusiophoresis (concentration of chemical field), thermophoresis (temperature field) etc [@anderson1989colloid]. A particle is "active" if it creates the phoretic field itself (for example, by a built-in chemical asymmetry on its surface), and the phenomena is called self-phoresis [@ebbens2010pursuit; @zhang2017active]. 
+The PyStokes library allows to study active particles in a Stokesian fluid [@kim2005]. Active particles are distinguished by their ability to produce flow, and thus motion, in the absence of external forces or torques on them. Examples of active particles include auto-phoretic synthetic particles [@ebbens2010pursuit] and microorganisms [@brennen1977]. An auto-phoretic particle self-propels due to non-equilibrium processes, such as chemical reactions, on its surface [@anderson1989colloid]. Examples of phoresis (phoretic field) are electrophoresis (electric field), diffusiophoresis (concentration of chemical field), thermophoresis (temperature field) etc [@anderson1989colloid]. A particle is "active" if it creates the phoretic field itself (for example, by a built-in chemical asymmetry on its surface), and the resulting phenomena is called self-phoresis [@ebbens2010pursuit; @zhang2017active]. 
+
+
+![Library structure and equations that determine the hydrodynamic and phoretic interactions between active particles in a three-dimensional domain $V$. The equations need to be solved with boundary conditions on the surface $S_{i}$ of the particles. To solve Laplace equation, the library takes input a set of expansion coefficients \begin{equation}\mathbf{J}_{i}^{(l)}\end{equation} for the prescribed active flux \begin{equation}j^{\mathcal A}\end{equation} on the surface of the $i$-th particle and computes the expansion coefficients of the surface concentration. The active slip velocity \begin{equation}\boldsymbol{v}^{\mathcal{A}}\end{equation} is obtained from this using the linear coupling relation \begin{equation}\boldsymbol{v}^{\mathcal{A}}=\mu_{c}\boldsymbol{\nabla}_{s}c\end{equation} [@anderson1989colloid]. The library outputs the expansion coefficients \begin{equation}\mathbf{V}_{i}^{(l\sigma)}\end{equation} of the active slip. To solve Stokes equation, the library takes as input these expansion coefficients, which may also be specified independently, and any body forces \begin{equation}\mathbf{F}_{i}^{B}\end{equation} and body torques \begin{equation}\mathbf{T}_{i}^{B}\end{equation} acting on the particles and returns their rigid body motion in terms of the velocities \begin{equation}\mathbf{V}_{i}\end{equation} and angular velocities \begin{equation}\boldsymbol{\Omega}_{i}\end{equation}.  Particle indices are i=1,\ldots,N and harmonic indices are l=1,2,\ldots and \sigma=s,a,t. \label{fig:figS}](FigSchema.png)
 
 Hydrodynamic and phoretic interactions between active particles in a viscous fluid are central to the understanding of their collective dynamics. 
 Under experimentally relevant conditions, the motion of the fluid is governed by the Stokes equation and that of the phoretic field, if one is present, by the Laplace equation. 
[…]
