+++ docs/paper/paper.bib
@@ -125,18 +125,6 @@
     year={2015}
 }
 
-@article{zhang2017active,
-    title={Active colloids with collective mobility status and research opportunities},
-    author={Zhang, J. and Luijten, E. and Grzybowski, B. A and Granick, S.},
-    journal={Chem. Soc. Rev.},
-    volume={46},
-    number={18},
-    pages={5551--5569},
-    year={2017},
-    publisher={Royal Society of Chemistry},
-    doi={10.1039/C7CS00461C},
-}
-
 @article{ebbens2010pursuit,
     title={In pursuit of propulsion at the nanoscale},
     author={Ebbens, S. J. and Howse, J. R.},
