+++ docs/paper/paper.bib
@@ -236,7 +236,7 @@
 @article{hinsen1995,
     title={HYDROLIB: a library for the evaluation of hydrodynamic interactions in colloidal suspensions},
     author={Hinsen, K.},
-    journal={Comp. Phys. Comm.},
+    journal={Comp. Phys. Commun.},
     volume={88},
     number={2},
     pages={327--340},
@@ -246,11 +246,11 @@
 
 @misc{libstokes,
     author={Ichiki, K.},
-    title={Py{R}itz},
+    title={libstokes},
     howpublished={\url{https://github.com/kichiki/libstokes}},
     publisher = {GitHub},
     journal = {GitHub repository},
-    year={2014},
+    year={2002},
 } 
 
 @book{kim2005,
