+++ docs/paper/paper.md
@@ -37,30 +37,6 @@ the force per unit area on the particle surfaces. This formally exact solution c
 truncated series expansion in a complete basis of functions on the particle boundaries. This, in turn, leads to an efficient and accurate numerical 
 method for computing hydrodynamic and phoretic interactions between active particles.
 
-
-
-# Methods
-
-\autoref{fig:figS} shows the the overall organization of the library. The expansion coefficients of the slip can be either specified (in the case of microorganisms) or obtained as a solution of Laplace equation (in the case of autophoretic particles [@singh2019competing]). Once the coefficients are determined, PyStokes solves the following equation numerically to obtain velocity and angular velocity of the $i$-th particle [@singh2015many; @singh2018generalized] 
-\begin{align}
-{\mathbf{V}}_{i}&=
-\boldsymbol{\mu}_{ij}^{TT}\cdot{\mathbf{F}_{j}^{P}}+ 
-\boldsymbol{\mu}_{ij}^{TR}\cdot{\mathbf{T}_{j}^{P}} + 
-\sum_{l\sigma=1s}^{\infty}\boldsymbol{\pi}_{ij}^{(T,\,l\sigma)}\cdot{\mathbf{V}_{j}^{(l\sigma)}},\qquad \boldsymbol{\mu}_{ij}^{\alpha\beta}\quad:\text{ mobility matrices},
-\\
-{\mathbf{\Omega}}_{i}&=\underbrace{
-\boldsymbol{\mu}_{ij}^{RT}\cdot\mathbf{F}_{j}^{P}+
-\boldsymbol{\mu}_{ij}^{RR}\cdot\mathbf{T}_{j}^{P}}_{\text{Passive}}+
-\sum_{l\sigma=1s}^{\infty}\underbrace{
-\boldsymbol{\pi}_{ij}^{(R,\,l\sigma)}\cdot\mathbf{{\mathbf{V}}}_{j}^{(l\sigma)}}_{\text{Active}},\qquad \boldsymbol{\pi}_{ij}^{(\alpha, l\sigma)}:\text{ propulsion tensors}.
-\end{align}
-In the above, sum over the repeated particle indices $j$ is implied and
-\begin{equation}
-\mathbf{F}_{i}^{P}:\text{ body force},\quad \mathbf{T}_{i}^{P}:\text{ body torque},\quad \mathbf{V}_{i}^{(l\sigma)}:\text{ expansion coefficients of active slip}.
-\end{equation}
-Thus, hydrodynamic interactions between particles with no-slip boundary conditions can be computed entirely in terms of mobility matrices [@kim2005]. There exists numerical libraries [@hinsen1995; @libstokes] to study such particles. The active contributions due to the slip boundary condition is given in terms of propulsion tensors [@singh2015many]. To the best of our knowledge, PyStokes is the only numerical implementation of propulsion matrices to model hydrodynamic interactions of active particles.   
-
-
 The PyStokes library can be used to compute the hydrodynamically interacting motion of squirming particles where the slip can be specified independently of a phoretic field, or the dynamics of passive suspensions where the slip vanishes and forces and torques are prescribed. PyStokes library can also compute Brownian motion, and thus, allows to study the interplay of passive, active and Brownian contributions. The PyStokes library has been used to model suspensions of microorganisms [@bolitho2020; @singh2016crystallization], 
 synthetic autophoretic particles [@singh2016crystallization; @singh2019competing] and 
 self-propelling droplets [@thutupalli2018FIPS].
@@ -68,6 +44,8 @@ Our software implementation uses a polylgot programming approach that combines t
 dynamically typed, interpreted language without sacrificing performance.
 
 
+
+# Methods
 
[…]
