+++ docs/paper/paper.bib
@@ -253,14 +253,6 @@
     year={2002},
 } 
 
-@book{kim2005,
-    title={Microhydrodynamics: Principles and Selected Applications},
-    author={Kim, S. and Karrila, S. J.},
-    publisher={Butterworth-Heinemann, Boston},
-    doi={10.1002/cite.330641004},
-    year={1991}
-}
-
 @article{brennen1977,
     author = "Brennen, C. and Winet, H.",
     doi = "10.1146/annurev.fl.09.010177.002011",
