+++ docs/paper/paper.md
@@ -47,7 +47,7 @@ dynamically typed, interpreted language without sacrificing performance.
 
 # Methods
 
-Our method relies on the reduction of linear elliptic partial differential equations to systems of linear algebraic equations using following steps.  
+Our method relies on the reduction of linear elliptic partial differential equations to systems of linear algebraic equations using following steps:  
 
 ![Key mathematical steps underpinning the PyStokes codebase.\label{fig:example}](figure.png)
 
@@ -68,10 +68,10 @@ the linear system can be evaluated analytically in terms of the Green's function
 infinite-dimensional linear system to a finite-dimensional one that can be solved by standard methods of linear algebra adapted for self-adjoint systems [@saad2003iterative]. 
 Analytical solution can be obtained by Jacobi iteration, which is equivalent to Smoluchowski's method of reflections. Numerical solutions can be obtained by the conjugate 
 gradient method, at a cost quadratic in the number of unknowns. From this solution, we can reconstruct the field and the flux on the boundary, use these to determine the 
-fields in the bulk, and from there, compute derived quantities. These steps have been elaborated in several 
-papers [@singh2015many; @singh2016crystallization; @singh2017fluctuation; @singh2018generalized; @singh2019competing; @bolitho2020] and we do not repeat them in detail here. 
+fields in the bulk, and from there, compute derived quantities. 
 
-\autoref{fig:figS} shows the the overall organization of the library. The expansion coefficients of the slip can be either specified (in the case of microorganisms) or obtained as a solution of Laplace equation (in the case of autophoretic particles [@singh2019competing]). Once the coefficients are determined, PyStokes solves the following equation numerically to obtain velocity and angular velocity of the $i$-th particle [@singh2015many; @singh2018generalized] 
+The above steps have been elaborated in several 
+papers [@singh2015many; @singh2016crystallization; @singh2017fluctuation; @singh2018generalized; @singh2019competing; @bolitho2020] and we do not repeat them in detail here. Briefly, the expansion coefficients of the slip can be either specified (microorganisms) or obtained as a solution of Laplace equation (autophoretic particles). Once the coefficients are determined, PyStokes solves the following equation numerically to obtain velocity and angular velocity of the $i$-th particle
 \begin{align}
 {\mathbf{V}}_{i}&=
 \boldsymbol{\mu}_{ij}^{TT}\cdot{\mathbf{F}_{j}^{P}}+ 
@@ -84,14 +84,14 @@ papers [@singh2015many; @singh2016crystallization; @singh2017fluctuation; @singh
 \sum_{l\sigma=1s}^{\infty}\underbrace{
 \boldsymbol{\pi}_{ij}^{(R,\,l\sigma)}\cdot\mathbf{{\mathbf{V}}}_{j}^{(l\sigma)}}_{\text{Active}},\qquad \boldsymbol{\pi}_{ij}^{(\alpha, l\sigma)}:\text{ propulsion tensors}.
 \end{align}
-In the above, sum over the repeated particle indices $j$ is implied and
+In the above, $\alpha,\beta=(T,R)$, repeated particle indices $j$ is summed, and
 \begin{equation}
 \mathbf{F}_{i}^{P}:\text{ body force},\quad \mathbf{T}_{i}^{P}:\text{ body torque},\quad \mathbf{V}_{i}^{(l\sigma)}:\text{ expansion coefficients of active slip}.
 \end{equation}
[…]
