+++ docs/paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'PyStokes: Phoresis and Stokesian hydrodynamics in Python'
+title: 'PyStokes: phoresis and Stokesian hydrodynamics in Python'
 tags:
   - Python
   - Cython
@@ -22,28 +22,18 @@ bibliography: paper.bib
 
 # Summary
 
-The PyStokes library is for studying hydrodynamic and phoretic interactions of active particles in Python. Active particles are distinguished by their ability to produce flow, and thus motion, in the absence of external forces or torques on them. Examples of active particles include auto-phoretic synthetic particles [@ebbens2010pursuit] and microorganisms [@brennen1977]. An auto-phoretic particle self-propels due to non-equilibrium processes, such as chemical reactions, on its surface [@anderson1989colloid]. Examples of phoresis (phoretic field) are electrophoresis (electric field), diffusiophoresis (concentration of chemical field), thermophoresis (temperature field) etc [@anderson1989colloid]. A particle is "active" if it creates the phoretic field itself (for example, by a built-in chemical asymmetry on its surface), and the resulting phenomena is called self-phoresis [@ebbens2010pursuit]. 
-
+PyStokes is a Python library for studying phoretic and hydrodynamic interactions between spherical particles when these interactions can be described by the solutions of, respectively, the Laplace and Stokes equations. The library has been specifically designed for studying these interactions in suspensions of active particles,  which are distinguished by their ability to produce flow, and thus motion, in the absence of external forces or torques. Such particles are endowed with a mechanism to produce hydrodynamic flow in a thin interfacial layer, which may be due to the motion of cilia, as in microogranisms [@brennen1977] or osmotic flows of various kinds in response to spontaneously generated gradients of phoretic fields [@ebbens2010pursuit]. The latter, often called autophoresis,  is a generalisation of well-known phoretic phenomena including, *inter alia*,  electrophoresis (electric field), diffusiophoresis (chemical field) and thermophoresis (temperature field) that occur in response to externally imposed gradients of phoretic fields [@anderson1989colloid]. 
 
 ![Input and output structure of PyStokes to determine the hydrodynamic and phoretic interactions between active particles in a three-dimensional domain $V$. The equations are coupled by active boundary conditions on the surface $S_{i}$ of the particles. Particle indices are $i=1,\ldots,N$ and harmonic indices are $l=1,2,\ldots$ and $\sigma=s,a,t$ (see text). \label{fig:figS}](FigSchema.png)
 
-Hydrodynamic and phoretic interactions between active particles in a viscous fluid are central to the understanding of their collective dynamics. 
-Under experimentally relevant conditions, the motion of the fluid is governed by the Stokes equation and that of the phoretic field, if one is present, by the Laplace equation. 
-The “activity” appears in these equations as boundary conditions on the particle surfaces that prescribe the slip velocity in the Stokes equation and flux of the phoretic field in the Laplace equation (see \autoref{fig:figS}). 
-The slip velocity and the phoretic flux are related by a linear constitutive law that can be derived from a detailed analysis of the boundary layer physics [@anderson1989colloid]. 
-The Stokes and Laplace equations are coupled by this linear constitutive law only at the particle boundaries. 
-The linearity of the governing equations and of the coupling boundary conditions allows for a formally exact solution of the problem of determining 
-the force per unit area on the particle surfaces. This formally exact solution can be approximated to any desired degree of accuracy by a 
[…]
