+++ pystokes/interface.pyx
@@ -609,7 +609,7 @@ cdef class Rbm:
 
 
     ## Noise
-    cpdef calcNoiseMuTT(self, double [:] v, double [:] r):
+    cpdef noiseTT(self, double [:] v, double [:] r):
         """
         Compute translation Brownian motion 
         ...
@@ -713,6 +713,106 @@ cdef class Rbm:
         return
 
 
+    cpdef noiseRR(self, double [:] o, double [:] r):
+        """
+        Compute rotational Brownian motion 
+        ...
+
+        Parameters
+        ----------
+        o: np.array
+            An array of angular velocities
+            An array of size 3*Np,
+        r: np.array
+            An array of positions
+            An array of size 3*Np,
+        """
+
+        cdef int i, j, Np=self.Np, xx=2*Np
+        cdef double dx, dy, dz, idr, h2, hsq, idr2, idr3, idr4, idr5
+        cdef double mur=1/(8*np.pi*self.eta), mu1=0.25*sqrt(2.0)*mur, mm=4/(self.a**3)
+        cdef double ox, oy, oz
+
+        cdef double [:, :] M = self.Mobility
+        cdef double [:]   Tr = np.random.normal(size=3*Np)
+
+
+        for i in prange(Np, nogil=True):
+            for j in range(Np):
+                dx = r[i]    - r[j]
+                dy = r[i+Np] - r[j+Np]
+                h2=2*r[j+xx]; hsq=r[j+xx]*r[j+xx]
+                if i!=j:
+                    dz = r[i+xx] - r[j+xx]
+                    idr = 1.0/sqrt( dx*dx + dy*dy + dz*dz )
+                    idr2=idr*idr;  idr3=idr*idr*idr
+                    dx = dx*idr; dy=dy*idr; dz=dz*idr
+                    #
+                    M[i,    j   ] = (2 - 6*dx*dx)*idr3
+                    M[i+Np, j+Np] = (2 - 6*dy*dy)*idr3
+                    M[i+xx, j+xx] = (2 - 6*dz*dz)*idr3
+                    M[i,    j+Np] = (  - 6*dx*dy)*idr3
+                    M[i,    j+xx] = (  - 6*dx*dz)*idr3
+                    M[i+Np, j+xx] = (  - 6*dy*dz)*idr3
+
+                    ###contributions from the image
+                    dz = r[i+xx] + r[j+xx]
+                    idr = 1.0/sqrt( dx*dx + dy*dy + dz*dz )
+                    dx = dx*idr; dy=dy*idr; dz=dz*idr
+                    idr2=idr*idr;  idr3=idr2*idr;  idr4=idr3*idr
+
+                    M[i,    j   ] += -(2 - 6*dx*dx)*idr3
+                    M[i+Np, j+Np] += -(2 - 6*dy*dy)*idr3
+                    M[i+xx, j+xx] += -(2 - 6*dz*dz)*idr3
+                    M[i,    j+Np] += -(  - 6*dx*dy)*idr3
+                    M[i,    j+xx] += -(  - 6*dx*dz)*idr3
+                    M[i+Np, j+xx] += -(  - 6*dy*dz)*idr3
+                else:
+                    # one-body mobility
+                    M[i,    j   ] = mm
+                    M[i+Np, j+Np] = mm
+                    M[i+xx, j+xx] = mm
+                    M[i,    j+Np] = 0
+                    M[i,    j+xx] = 0
+                    M[i+Np, j+xx] = 0
+
[…]
