+++ docs/sphinx/README.md
@@ -1,0 +1,7 @@
+Once all the .rst files are updated following [autodoc](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html), run the following
+
+```bash
+pip install -U Sphinx
+pip install sphinx_rtd_theme
+make html
+```
