### 2020-06-29 #2318 @rajeshrinet
@harpolea: thank you very much for your editorial assistance in reviewing our [submission](https://github.com/rajeshrinet/pystokes/blob/master/docs/paper/paper.md). We, the authors (@rajeshrinet and @ronojoy), have revised the manuscript in response to the reviewers suggestions. We detail the changes below and hope that, with these changes, the manuscript can now accepted in JOSS.

Referee A: @khinsen.

> @harpolea I did a first pass looking at the code and the paper. It was a lot of fun: this is a very nice implementation of hydrodynamic and phoretic interactions which invites users to play with it, in particular by including visualization tools, and that is just great for doing research. The next time I do work in this domain, I'll happily abandon [my own 25-year-old Fortran library](https://github.com/khinsen/HYDROLIB) and use PyStokes instead!
>
> I have made a few suggestions for improvements via issues/pull requests on the code repository, which the authors have handled immediately. Interactive reviewing is a fun experience :-)

We thank the referee for a positive evaluation of our code and paper. It has indeed been fun to interactively improve the codebase through discussion on the issue tracker. Responses to specific comments are below:

>
> There are a few open points in the review checklist as of today:
>
> 1. There are no tests, only examples. This being number-crunching code, it is very difficult to write tests that are meaningful and predictable, given how little control programmers have over the details of float arithmetic (shameless self-advertising: [The Approximation Tower in Computational Science: Why Testing Scientific Software Is Difficult](https://hal.archives-ouvertes.fr/hal-01171382)). Does JOSS have an editorial policy concerning tests for number-crunching software? Or at least recommendations for authors? I didn't find anything.

We have now upgraded our testing and made it more visible. We followed the referee's suggestion to add details to the REAMDE on how to test the code.

> 2. Community guidelines: There are instructions for reporting issues and for seeking support, but not for contributing to the software. See [rajeshrinet/pystokes#6](https://github.com/rajeshrinet/pystokes/issues/6).

A [CONTRIBUTING.md](https://github.com/rajeshrinet/pystokes/blob/master/CONTRIBUTING.md) was added.

> 3. High-level summary in the paper: The current summary is clear enough for potential users of the software, but maybe not "for a diverse, non-specialist audience". But this should best be judged by someone unfamiliar with the field.

We have now added a new paragraph at the start of the paper and added a new figure. Both are at page 1. We hope that this will be helpful as a high-level summary.

> 4. State of the field: the paper doesn't provide comparisons with other commonly-used packages, but then there are no competing commonly used packages, as far as I can tell. But the authors should definitely say something about this, even if it's only a confirmation of my suspicion.

We have now clearly described on page 3 of our paper that hydrodynamic interactions of active particles are obtained in terms of mobility matrices and propulsion tensors. Two libraries which compute mobility matrices have been cited, while we are not aware of any library, other than PyStokes, that evaluates propulsion tensors to compute hydrodynamic interactions of active particles.

Referee 2: @fergusc

> I think the software looks very good.
>
> I found a few problems with installation, but @rajeshrinet addressed those very quickly - it's great for this to be happening in real time!
>
> In terms of the formal checklist, I have the same two checkboxes un-checked as @khinsen:
>

We thank the referee for a positive evaluation and comments on our submission. It was indeed very helpful in improving installation instructions using PyPI.

> * **Summary**: I have some experience in this area (I once coded up a spectral method for solving Navier-Stokes on a periodic domain), but I am not a domain expert and do not understand the whole of the first paragraph.  Even after a bit of time on Google I'm still not certain I know what the phoretic field is, for instance.  I think a "diverse non-specialist audience" would benefit from a paragraph up-front with a higher level overview of what the software is designed to do.
> * **State of the field**: I cannot tell from reading the paper whether there is other software available that does a similar task.  Either way, I think this needs stating in the paper.
>

As mentioned in response to referee A (@khinsen), we have now revised the paper and added more details for a non-specialist audience including a figure on page 1 which gives a high-level summary.

> Other than that, a couple of suggestions:
>
> * The README (until very recently) stated that PyStokes works with Python 2.6+ or Python 3.4+: I've not been able to verify this, but it would be very useful for users to know which versions it works with.  It would be very easy to add a GitHub workflow to test on multiple Python version.  @rajeshrinet would you be interested in a PR that implements that?

> * Testing (outside of checking that examples run - which is great!) is fairly thin.  I would encourage coverage checking (suggested [here](https://github.com/rajeshrinet/pystokes/issues/10) to assess what proportion of code is covered by tests, but on top of that it would be great to see a unit testing framework in use.  I understand that some of the number crunching can be very hard to test effectively: but even boring things like the utils would benefit from some testing.

As discussed [here](https://github.com/rajeshrinet/pystokes/issues/10), coverage checking for a cython codebase leads to performance loss. Thus, we have decided against coverage checking, pending a solution that does not compromise on speed.
