+++ pystokes/mima.pyx
@@ -1,913 +1,0 @@
-cimport cython
-from libc.math cimport sqrt, exp, pow, erfc, sin, cos
-from cython.parallel import prange
-import numpy as np
-cimport numpy as np
-cdef double PI = 3.14159265359
-
-cdef extern from "complex.h" nogil:
-    double complex cexp(double complex x)
-
-@cython.boundscheck(False)
-@cython.cdivision(True)
-@cython.nonecheck(False)
-@cython.wraparound(False)
-cdef class Mima3D:
-    """
-    Mollified irreducible multipole approach (MIMA) of a 3D Stokes flow
-    MIMA resolves the fluid flow and solves Stokes equation directly.
-    For further details of this method, read chapter 10 of the thesis: 
-    https://dspace.imsc.res.in/xmlui/handle/123456789/418 
-    or at https://www.imsc.res.in/xmlui/handle/123456789/418
-
-    """
-
-    def __init__(self, a_, Np_, Lx_, Ly_, Lz_, Nx_, Ny_, Nz_ ):
-        self.a = a_
-        self.Np = Np_
-        self.Lx = Lx_
-        self.Ly = Ly_
-        self.Lz = Lz_
-        self.Nx = Nx_
-        self.Ny = Ny_
-        self.Nz = Nz_
-
-        self.facx = 2*PI/Nx_
-        self.facy = 2*PI/Ny_
-        self.facz = 2*PI/Nz_
-        
-        self.fx0 = np.zeros((Nx_, Ny_, Nz_), dtype=np.float) 
-        self.fk0 = np.zeros((Nx_, Ny_, Nz_), dtype=np.complex128)
-        self.fkx = np.empty((Nx_, Ny_, Nz_), dtype=np.complex128) 
-        self.fky = np.empty((Nx_, Ny_, Nz_), dtype=np.complex128)
-        self.fkz = np.empty((Nx_, Ny_, Nz_), dtype=np.complex128)
-        self.vkx = np.empty((Nx_, Ny_, Nz_), dtype=np.complex128) 
-        self.vky = np.empty((Nx_, Ny_, Nz_), dtype=np.complex128)
-        self.vkz = np.empty((Nx_, Ny_, Nz_), dtype=np.complex128)
-       
-
-    cpdef flowField1s(self, np.ndarray v, double [:] r, double [:] F, double sigma=3, int NN=20):
-        """
-        Compute flow field at field points  due to body forces 
-        ...
-
-        Parameters
-        ----------
-        v: np.array
-            An array of flow at field points
-            An array of size Nx x Ny x Nz,
-        r: np.array
-            An array of positions
-            An array of size Nx x Ny x Nz,
-        F: np.array
-            An array of body forces
-            An array of size 3*Np,
-        """
-        cdef:
-            int i, nx, ny, nz , x, y, z, Nx, Ny, Nz, Np, jx, jy, jz
-            double arg, facx, facy, facz, kx, ky, kz
-            double kdotdr, a2, k2
-        Nx, Ny, Nz = self.Nx, self.Ny, self.Nz 
-        facx, facy, facz = self.facx, self.facy, self.facz
-        Np, a2 = self.Np, self.a*self.a/6.0
-        
-        self.fourierFk(sigma, NN)  # construct the fk0 
-            
-        cdef complex [:, :, :] fk0 = self.fk0
-        cdef complex [:, :, :] fkx = self.fkx
-        cdef complex [:, :, :] fky = self.fky
-        cdef complex [:, :, :] fkz = self.fkz
-
-        for jz in prange(Nz, nogil=True): 
-            for jy in range(Ny): 
[…]
