+++ examples/others/mima/README.md
@@ -1,5 +1,0 @@
-## MIMA in pystokes
-Mollified irreducible multipole approach (MIMA) resolves the fluid flow and solves Stokes equation directly.  
-
-
-The .ipynb files illustrate the usage. 
