+++ .github/workflows/build-test-win.yml
@@ -1,0 +1,31 @@
+name: build-win
+
+on: [push, pull_request]
+
+jobs:
+
+  build-and-test:
+    name: install and test win
+    runs-on: windows-latest
+    strategy:
+      matrix:
+        python-version: [2.7, 3.7]
+    steps:
+      - uses: actions/checkout@v2
+    
+      - name: Setup Miniconda
+        uses: goanpeca/setup-miniconda@v1.2.0
+        with:
+         activate-environment: pystokes
+         environment-file: environment.yml
+         python-version: ${{ matrix.python-version }}
+      - name: make-pystokes
+        shell: bash -l {0}
+        run: |
+          conda activate pystokes
+          make
+      - name: test-pystokes
+        shell: bash -l {0}
+        run: |
+          make test
+
