+++ docs/sphinx/mima.rst
@@ -1,16 +1,0 @@
-Mollified irreducible multipole approach 
-========================================
-
-Mollified irreducible multipole approach (MIMA) to Stokesian hydrodynamics
-
-Mima3D
-----------------------------------------
-.. autoclass:: pystokes.mima.Mima3D
-    :members:
-
-
-Mima2D
---------------------------
-.. autoclass:: pystokes.mima.Mima2D
-    :members:
-
