+++ pystokes/__init__.py
@@ -1,5 +1,4 @@
 import pystokes.interface
-import pystokes.mima
 import pystokes.periodic
 import pystokes.twoWalls
 import pystokes.unbounded
