+++ .github/workflows/build-test-linux.yml
@@ -1,4 +1,4 @@
-name: build
+name: build-linux
 
 on: [push, pull_request]
 
