### 2020-06-29 #2318 @fcooper8472
The additional paragraph at the start really helps - it's clear and certainly makes it more accessible to a non-specialist.

The only remaining concern is still about other software packages.  I still do not think there is an assertion that no other similar software exists, or a description of other packages, so I'm still not 100% sure I can tick

"Do the authors describe how this software compares to other commonly-used packages?"

@rajeshrinet are there other packages that do anything in the same vein?  If not could you add a sentence to this effect?
