### 2020-06-29 #2318 @rajeshrinet
@fcooper8472 thanks for your comments.

A  search on GitHub does not find anything comparable to PyStokes. There are libraries, such as Ludwig (from EPCC, Edinburgh) which simulates the hydrodynamics of point particles, resolving the fluid degrees of freedom. It does not include phoretic interactions. In terms of the functionality of simultaneously resolving both phoretic and hydrodynamic interactions, we think that PyStokes is, as of this writing, unique. Thus we note in our revision that:

> To the best of our knowledge, PyStokes is the only numerical implementation of propulsion tensors to model suspensions of active particles.
