+++ .github/workflows/build-test.yml
@@ -1,15 +1,17 @@
-name: build-windows
+name: build
 
 on: [push, pull_request]
 
 jobs:
 
   build-and-test:
-    name: install and test win
-    runs-on: windows-latest
+    name: build (${{ matrix.os }})
+    runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
-        python-version: [2.7, 3.7]
+        os: ["ubuntu-latest", "macos-latest","windows-latest"]
+        python-version: ["3.7"]
     steps:
       - uses: actions/checkout@v2
     
