### 2020-06-30 #2318 @khinsen
@harpolea @rajeshrinet I am happy with the updated paper as well. Figure 1 in particular is a very nice summary of how everything fits together. The authors' description of the state of the art looks accurate as well. So... I just ticked off the last boxes of my checklist! Ready for take-off!
