### 2020-06-30 #2318 @khinsen
Ahhh.... a minor nitpick I noticed while closing the open PDF file! The proper title for Odqvist 1930 is "Über die Randwertaufgaben der Hydrodynamik zäher Flüssigkeiten", with all capitalization required by German grammar.
