+++ docs/paper/paper.bib
@@ -13,6 +13,7 @@
     volume={19},
     number={9},
     pages={735--748},
+    doi={10.1137/1.9781611973242},
     year={1966}
 }
 
@@ -62,7 +63,7 @@
 }
 
 @article{fkg1930bandwertaufgaben,
-    title={{\"U}ber die Bandwertaufgaben der Hydrodynamik z{\"a}her Fl{\"u}ssig-keiten},
+    title={{\"U}ber die {B}andwertaufgaben der {H}ydrodynamik z{\"a}her {F}l{\"u}ssig-keiten},
     author={Odqvist, F. K. G.},
     journal={Mathematische Zeitschrift},
     volume={32},
@@ -241,12 +242,13 @@
     number={2},
     pages={327--340},
     year={1995},
+    doi={10.1016/0010-4655(95)00029-F},
     publisher={Elsevier}
 }
 
 @misc{libstokes,
     author={Ichiki, K.},
-    title={libstokes},
+    title={{l}ibstokes},
     howpublished={\url{https://github.com/kichiki/libstokes}},
     publisher = {GitHub},
     journal = {GitHub repository},
