### 2020-06-30 #2318 @harpolea
Thanks @khinsen!

@rajeshrinet: I'm just going through the references now. As whedon pointed out, there are a couple that are missing DOIs (the first one doesn't seem to match up, but the latter two look good to me). Would you be able to add them?
