### 2020-06-30 6a79008
Fix references in paper.bib
I've made a couple of changes to paper.bib to fix the references:

Added a title to bolitho2020.

Removed the incorrect DOI for finlayson1966method and instead provided a url (I was unable to find a correct DOI)
