+++ docs/paper/paper.bib
@@ -13,7 +13,7 @@
     volume={19},
     number={9},
     pages={735--748},
-    doi={10.1137/1.9781611973242},
+    url={http://faculty.washington.edu/finlayso/MWR-AReview.pdf},
     year={1966}
 }
 
@@ -151,6 +151,7 @@
 }
 
 @article{bolitho2020,
+    title = {Periodic Orbits of Active Particles Induced by Hydrodynamic Monopoles},
     author = {Bolitho, A. and Singh, R. and Adhikari, R.},
     journal = {Phys. Rev. Lett.},
     doi={10.1103/PhysRevLett.124.088003},
