+++ docs/paper/paper.bib
@@ -100,6 +100,7 @@
     volume = {26},
     number = {1-2},
     pages = {78-97},
+    doi={10.1080/17797179.2017.1294829},
     url={http://www.tandfonline.com/doi/pdf/10.1080/17797179.2017.1294829},
     year={2017}
 }
@@ -122,6 +123,7 @@
     volume={2015},
     number={6},
     pages={P06017},
+    doi={10.1088/1742-5468/2015/06/p06017},
     url={http://stacks.iop.org/1742-5468/2015/i=6/a=P06017},
     year={2015}
 }
