### 2020-06-30 #2318 @harpolea
@rajeshrinet thanks! I've made a PR with fixes to a couple of the references. One was missing a title and for another the DOI pointed to a different article
