### 2020-06-30 #2318 @rajeshrinet
> @rajeshrinet thanks! I've made a PR with fixes to a couple of the references. One was missing a title, and for another the DOI pointed to a different article

Thanks @harpolea. It has now beed merged.
