### 2020-06-30 #2318 @rajeshrinet
@harpolea ,

* I have now added a new tagged release: [v2.2.0](https://github.com/rajeshrinet/pystokes/releases/tag/v2.2.0)

* the version has been archived on Zeonodo: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3923867.svg)](https://doi.org/10.5281/zenodo.3923867) <br/>
