### 2020-06-30 #2318 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
